% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-abc_prior-class.R
\name{abc_prior}
\alias{abc_prior}
\alias{new_abc_prior}
\alias{as.abc_prior}
\alias{is.abc_prior}
\alias{format.abc_prior}
\alias{print.abc_prior}
\alias{plot.abc_prior}
\title{\code{abc_prior} S3 class}
\usage{
new_abc_prior(.dists, .constraints = list(), .derived = list(), .cor = NULL)

as.abc_prior(x, ...)

is.abc_prior(x, ...)

\method{format}{abc_prior}(x, ...)

\method{print}{abc_prior}(x, ...)

\method{plot}{abc_prior}(x, ...)
}
\arguments{
\item{.dists}{distribution functions as a named list of S3 \code{dist_fns} objects}

\item{.constraints}{a list of one sided formulae the result each of which should
evaluate to a boolean when compared against the names of the priors and
derived values.}

\item{.derived}{a list of two sided formulae. The RHS refer to the
priors, and the LHS as a name to derive.}

\item{.cor}{(optional) a correlation matrix for the priors}

\item{x}{an \code{abc_prior} S3 object}

\item{...}{passed on to methods}
}
\value{
an S3 object of class \code{abc_prior} which contains
\itemize{
\item a list of \code{dist_fns}
\item a \code{cor} attribute describing their correlation
\item a \code{derived} attribute describing derive values
\item a \code{constraints} attribute listing the constraints
\item a \code{params} attribute listing the names of the parameters
}
}
\description{
\code{abc_prior} S3 class
}
\section{Methods (by generic)}{
\itemize{
\item \code{format(abc_prior)}: Format an \code{abc_prior}

\item \code{print(abc_prior)}: Print an \code{abc_prior}

\item \code{plot(abc_prior)}: Plot an \code{abc_prior}

}}
\section{Functions}{
\itemize{
\item \code{new_abc_prior()}: Create a new prior

\item \code{as.abc_prior()}: Create a prior from a named list of \code{dist_fns}

\item \code{is.abc_prior()}: Test is an \code{abc_prior}

}}
\concept{abc_prior_s3}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

p = new_abc_prior(
  .dists = list(
    mean = as.dist_fns("norm",4,2),
    sd = as.dist_fns("gamma",2)
  ),
  .derived = list(
    shape ~ mean^2 / sd^2,
    rate ~ mean / sd^2
  ),
  .constraints = list(
    ~ mean > sd
  )
)

testthat::expect_equal(
  format(p),
  "Parameters: \n* mean: norm(mean = 4, sd = 2)\n* sd: gamma(shape = 2, rate = 1)\nConstraints:\n* mean > sd\nDerived values:\n* shape = mean^2/sd^2\n* rate = mean/sd^2"
)

}\if{html}{\out{</div>}}
}

