% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{as.character.dist_fns}
\alias{as.character.dist_fns}
\title{Cast a \code{dist_fns} S3 object to a character vector}
\usage{
\method{as.character}{dist_fns}(x, ...)
}
\arguments{
\item{x}{a \code{dist_fns} S3 object}

\item{...}{passed onto methods}
}
\value{
a character vector
}
\description{
Boilerplate S3 class operation
}
\concept{dist_fns_s3}
\keyword{internal}
