% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{as.character.link_fns_list}
\alias{as.character.link_fns_list}
\title{Convert a \code{link_fns_list} to character}
\usage{
\method{as.character}{link_fns_list}(x, ...)
}
\arguments{
\item{x}{a \code{link_fns_list} S3 object}

\item{...}{passed onto methods}
}
\value{
a character vector
}
\description{
These boilerplate functions allow generic list behaviour from \code{link_fns_list}
classes allowing \code{link_fns} S3 objects to be used in lists or dataframes.
}
\concept{link_fns_s3}
\keyword{internal}
