% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{as.dist_fns.default}
\alias{as.dist_fns.default}
\title{Default method throws error}
\usage{
\method{as.dist_fns}{default}(x, ...)
}
\arguments{
\item{x}{a \code{dist_fns} S3 object}

\item{...}{passed onto methods}
}
\value{
nothing
}
\description{
Boilerplate S3 class operation
}
\concept{dist_fns_s3}
\keyword{internal}
