% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{as.dist_fns.character}
\alias{as.dist_fns.character}
\alias{as.dist_fns.function}
\alias{as.dist_fns.fitdist}
\alias{as.dist_fns}
\title{Create a \code{dist_fns} S3 object}
\usage{
\method{as.dist_fns}{character}(x, ...)

\method{as.dist_fns}{`function`}(x, ...)

\method{as.dist_fns}{fitdist}(x, ...)

as.dist_fns(x, ...)
}
\arguments{
\item{x}{a \code{dist_fns} S3 object}

\item{...}{passed onto methods}
}
\value{
a \code{dist_fns} S3 object
}
\description{
A class wrapping a single (or set) of parametrised distributions and
provides access to the quantile, cumulative probability and random functions
of that specific distribution. Parametrisation is handled on construction.
}
\details{
\code{dist_fns} and \code{dist_fns_list} objects support \code{$} access for fields and \code{@} access
for attributes. \code{dist_fns_list}s can be made with the \code{c()} or \code{rep()} functions,
or with the \code{purrr} style map functions, and they support subsetting.
Individual \code{dist_fns} members of \code{dist_fns_list}s can be accessed with \code{[[}.
}
\section{Methods (by class)}{
\itemize{
\item \code{as.dist_fns(character)}: Construct a distribution by name

\item \code{as.dist_fns(`function`)}: From a statistical function

\item \code{as.dist_fns(fitdist)}: From a \code{fitdistrplus::fitdist} output

}}
\concept{dist_fns_s3}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

pois = as.dist_fns("pois",lambda = 8)
n = as.dist_fns("norm",mean=4)
}\if{html}{\out{</div>}}
}

