% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{as.dist_fns_list}
\alias{as.dist_fns_list}
\title{Cast to a list of \code{dist_fns} S3 objects}
\usage{
as.dist_fns_list(x)
}
\arguments{
\item{x}{a \code{dist_fns_list} S3 object}
}
\value{
a \code{dist_fns_list} S3 object
}
\description{
This function wraps \code{dist_fns} and unwraps plain lists such that
the result is a flat \code{dist_fns_list} containing \code{dist_fns} objects only
}
\concept{dist_fns_s3}
\keyword{internal}
