% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{as.link_fns.default}
\alias{as.link_fns.default}
\title{Default method throws error}
\usage{
\method{as.link_fns}{default}(x, ...)
}
\arguments{
\item{x}{a \code{link_fns} S3 object}

\item{...}{passed onto methods}
}
\value{
nothing
}
\description{
Boilerplate S3 class operation
}
\concept{link_fns_s3}
\keyword{internal}
