% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{as.link_fns.character}
\alias{as.link_fns.character}
\alias{as.link_fns.dist_fns}
\alias{as.link_fns.family}
\alias{as.link_fns}
\title{Create a \code{link_fns} S3 object}
\usage{
\method{as.link_fns}{character}(x, ...)

\method{as.link_fns}{dist_fns}(x, ...)

\method{as.link_fns}{family}(x, ...)

as.link_fns(x, ...)
}
\arguments{
\item{x}{a statistical family as a \code{stats::family} S3 object}

\item{...}{ignored}
}
\value{
a \code{link_fns} S3 object
}
\description{
The link function class allows forwards and backwards transformation.
Link functions can be defined by name or using a statistical distribution
in which case the forward link is a logit of the cumulative probability
and the reverse is the quantile of the expit.
}
\details{
\code{link_fns} and \code{link_fns_list} objects support \code{$} access for fields and \code{@} access
for attributes. \code{link_fns_list}s can be made with the \code{c()} or \code{rep()} functions,
or with the \code{purrr} style map functions, and they support subsetting.
Individual \code{link_fns} members of \code{link_fns_list}s can be accessed with \code{[[}.
}
\section{Methods (by class)}{
\itemize{
\item \code{as.link_fns(character)}: Link function from name

\item \code{as.link_fns(dist_fns)}: Link function from name

\item \code{as.link_fns(family)}: Link function from name

}}
\concept{link_fns_s3}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

links = c("ident", "log", "logit", "probit", "cloglog", "neginv", "inv2")
test = seq(0.1,0.9,0.1) # within support of all links
for (l in links) {
  lfn = as.link_fns(l)
  t = lfn$trans(test)
  i = lfn$inv(t)
  testthat::expect_equal(i,test)
}
}\if{html}{\out{</div>}}
}

