% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{as.list.dist_fns}
\alias{as.list.dist_fns}
\title{Convert a \code{dist_fns} S3 object into a plain list}
\usage{
\method{as.list}{dist_fns}(x, ...)
}
\arguments{
\item{x}{a \code{dist_fns} S3 object}

\item{...}{passed onto methods}
}
\value{
the internal structure of the object as a plain list
}
\description{
Boilerplate S3 class operation
}
\concept{dist_fns_s3}
\keyword{internal}
