% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{as.list.link_fns_list}
\alias{as.list.link_fns_list}
\title{Un format the \code{link_fns_list}}
\usage{
\method{as.list}{link_fns_list}(x, ...)
}
\arguments{
\item{x}{a \code{link_fns_list} S3 object}

\item{...}{passed onto methods}
}
\description{
These boilerplate functions allow generic list behaviour from \code{link_fns_list}
classes allowing \code{link_fns} S3 objects to be used in lists or dataframes.
}
\concept{link_fns_s3}
\keyword{S3}
\keyword{`link_fns`}
\keyword{a}
\keyword{internal}
\keyword{list}
\keyword{objects}
\keyword{of}
\keyword{plain}
