% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{@.dist_fns}
\alias{@.dist_fns}
\title{Extract named attribute from a \code{dist_fns}}
\usage{
\method{@}{dist_fns}(x, y)
}
\arguments{
\item{x}{a \code{dist_fns} S3 object}

\item{y}{item to retrieve}
}
\value{
an attribute value for \code{x}
}
\description{
Boilerplate S3 class operation
}
\concept{dist_fns_s3}
\keyword{internal}
