% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{@.link_fns}
\alias{@.link_fns}
\title{Extract named attribute from a \code{link_fns}}
\usage{
\method{@}{link_fns}(x, y)
}
\arguments{
\item{x}{a \code{link_fns} S3 object}

\item{y}{item to retrieve}
}
\value{
an attribute value for \code{x}
}
\description{
Boilerplate S3 class operation
}
\concept{link_fns_s3}
\keyword{internal}
