% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-crate.R
\name{at.crate}
\alias{at.crate}
\alias{$.crate}
\title{Explore crated data for a \code{crate}}
\usage{
\method{$}{crate}(x, y)
}
\arguments{
\item{x}{a crated function}

\item{y}{item to retrieve}
}
\value{
crated data from \code{x}
}
\description{
Explore crated data for a \code{crate}
}
\concept{crate_utils}
\keyword{internal}
