% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-abc_prior-class.R, R/s3-dist_fns-class.R
\name{at.dist_fns}
\alias{at.dist_fns}
\alias{@.abc_prior}
\alias{@.dist_fns}
\title{Extract named attribute from a \code{abc_prior}}
\usage{
\method{@}{abc_prior}(x, y)

\method{@}{dist_fns}(x, y)
}
\arguments{
\item{x}{an \code{abc_prior} S3 object}

\item{y}{item to retrieve}
}
\value{
an attribute value for \code{x}

an attribute value for \code{x}
}
\description{
Boilerplate S3 class operation
}
\concept{abc_prior_s3}
\concept{dist_fns_s3}
\keyword{internal}
