% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{c.dist_fns}
\alias{c.dist_fns}
\title{Concatenate a \code{dist_fns} S3 object or \code{dist_fns_list}s}
\usage{
\method{c}{dist_fns}(...)
}
\arguments{
\item{...}{some of \code{dist_fns}, \code{dist_fns_list}s or lists of \code{dist_fns}}
}
\value{
a \code{dist_fns_list}
}
\description{
Concatenate a \code{dist_fns} S3 object or \code{dist_fns_list}s
}
\concept{dist_fns_s3}
