% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{c.dist_fns_list}
\alias{c.dist_fns_list}
\title{Concatenate or construct a \code{dist_fns_list}}
\usage{
\method{c}{dist_fns_list}(...)
}
\arguments{
\item{...}{some of \code{dist_fns_list} and \code{dist_fns} or list of \code{dist_fns}s}
}
\value{
a flattened \code{dist_fns_list} S3 object
}
\description{
Concatenate or construct a \code{dist_fns_list}
}
\concept{dist_fns_s3}
\keyword{internal}
