% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{c.link_fns}
\alias{c.link_fns}
\title{Concatenate a \code{link_fns} S3 object or \code{link_fns_list}s}
\usage{
\method{c}{link_fns}(...)
}
\arguments{
\item{...}{some of \code{link_fns}, \code{link_fns_list}s or lists of \code{link_fns}}
}
\value{
a \code{link_fns_list}
}
\description{
Concatenate a \code{link_fns} S3 object or \code{link_fns_list}s
}
\concept{link_fns_s3}
