% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{c.link_fns_list}
\alias{c.link_fns_list}
\title{Concatenate or construct a \code{link_fns_list}}
\usage{
\method{c}{link_fns_list}(...)
}
\arguments{
\item{...}{some of \code{link_fns_list} and \code{link_fns} or list of \code{link_fns}s}
}
\value{
a flattened \code{link_fns_list} S3 object
}
\description{
Concatenate or construct a \code{link_fns_list}
}
\concept{link_fns_s3}
\keyword{internal}
