% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{$.link_fns_list}
\alias{$.link_fns_list}
\title{Extract named item(s) from a \code{link_fns_list}}
\usage{
\method{$}{link_fns_list}(x, y)
}
\arguments{
\item{x}{a \code{link_fns_list} S3 object}

\item{y}{item to retrieve}
}
\value{
a list or vector of the named items
}
\description{
These boilerplate functions allow generic list behaviour from \code{link_fns_list}
classes allowing \code{link_fns} S3 objects to be used in lists or dataframes.
}
\concept{link_fns_s3}
\keyword{internal}
