% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{dist_fns}
\alias{dist_fns}
\title{Create an empty \code{dist_fns_list}}
\usage{
dist_fns()
}
\value{
an empty \code{dist_fns_list}
}
\description{
Boilerplate S3 class operation
}
\concept{dist_fns_s3}
