% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{dlogitnorm}
\alias{dlogitnorm}
\title{Logit-normal distribution}
\usage{
dlogitnorm(x, meanlogit = 0, sdlogit = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles \verb{(0<x<1)}}

\item{meanlogit}{the mean on the logit scale}

\item{sdlogit}{the sd on the logit scale}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
a vector of probabilities, quantiles, densities or samples.
}
\description{
The logit-normal distribution has a support of 0 to 1.
}
\examples{
dlogitnorm(seq(0.1,0.9,0.1), 0, 1)
}
\concept{distributions}
