% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{dlogitnorm2}
\alias{dlogitnorm2}
\title{Logit-normal distribution}
\usage{
dlogitnorm2(x, prob.0.5 = 0.5, kappa = 1 - exp(-1), log = FALSE)
}
\arguments{
\item{x}{vector of quantiles \verb{(0<x<1)}}

\item{prob.0.5}{the median on the true scale}

\item{kappa}{a dispersion parameter from 0 (none) to 1 maximum dispersion}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
a vector of probabilities, quantiles, densities or samples.
}
\description{
The logit-normal distribution has a support of 0 to 1.
}
\examples{

q = seq(0.1,0.9,0.1)
eps = sqrt(.Machine$double.eps)
dlogitnorm2(q, 0.75, 0.2)

(plogitnorm2(q+eps, 0.75, 0.2) -
  plogitnorm2(q-eps, 0.75, 0.2)) / (2*eps)
}
\concept{distributions}
