% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{dnull}
\alias{dnull}
\title{Null distributions always returns NA}
\usage{
dnull(x, ...)
}
\arguments{
\item{x}{vector of quantiles}

\item{...}{not used}
}
\description{
Null distributions always returns NA
}
\examples{
dnull(c(0.25,0.5,0.75), 0.5, 0.25)
}
\concept{distributions}
