% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-abc_prior-class.R
\name{.AtNames.abc_prior}
\alias{.AtNames.abc_prior}
\title{Support for auto suggests on an \code{abc_prior}s}
\usage{
\method{.AtNames}{abc_prior}(x, pattern)
}
\arguments{
\item{x}{an \code{abc_prior} S3 object}

\item{pattern}{a regular expression}
}
\value{
the names of the attributes
}
\description{
Support for auto suggests on an \code{abc_prior}s
}
\concept{abc_prior_s3}
\keyword{internal}
