% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{.AtNames.dist_fns}
\alias{.AtNames.dist_fns}
\title{Support for auto suggests on \code{dist_fns_list}s}
\usage{
\method{.AtNames}{dist_fns}(x, pattern)
}
\arguments{
\item{x}{a \code{dist_fns} S3 object}

\item{pattern}{a regular expression}
}
\value{
the names of the attributes
}
\description{
Boilerplate S3 class operation
}
\concept{dist_fns_s3}
\keyword{internal}
