% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{.AtNames.dist_fns_list}
\alias{.AtNames.dist_fns_list}
\title{Support for auto suggests on \code{dist_fns_list}s}
\usage{
\method{.AtNames}{dist_fns_list}(x, pattern)
}
\arguments{
\item{x}{a \code{dist_fns_list} S3 object}
}
\value{
the names of the attributes
}
\description{
These boilerplate functions allow generic list behaviour from \code{dist_fns_list}
classes allowing \code{dist_fns} S3 objects to be used in lists or dataframes.
}
\concept{dist_fns_s3}
\keyword{internal}
