% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{.AtNames.link_fns}
\alias{.AtNames.link_fns}
\title{Support for auto suggests on \code{link_fns_list}s}
\usage{
\method{.AtNames}{link_fns}(x, pattern)
}
\arguments{
\item{x}{a \code{link_fns} S3 object}

\item{pattern}{a regular expression}
}
\value{
the names of the attributes
}
\description{
Boilerplate S3 class operation
}
\concept{link_fns_s3}
\keyword{internal}
