% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-crate.R
\name{.autocrate}
\alias{.autocrate}
\title{Crate a function that may not have all the namespaces qualified.}
\usage{
.autocrate(expr, ...)
}
\arguments{
\item{expr}{an expression defining a function}

\item{...}{named list of global variables}
}
\value{
a crated function
}
\description{
Regular crating requires everything be namespaced and which is easy to
forget, particularly for \verb{stats::} and \verb{dplyr::}.
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{


z = 13
tmp = .autocrate(function(x) {
  y <- runif(x)
  stats::rnorm(x,y)
  x[1] <- 12
  x[2] <- !!z # inlining happens during expression parsing...?
})

# unqualified reference is qualified
testthat::expect_equal(format(body(tmp)), c(
  "{",
  "    y <- stats::runif(x)",
  "    stats::rnorm(x, y)",
  "    x[1] <- 12",
  "    x[2] <- 13",
  "}"
))

tmpfn = function(x) {
  "temp"
}
tmpvar = 10

testthat::expect_error(
  {
    .autocrate(function(y) {
      if (tmpvar == 10) tmpfn(y)
    })
  },
  "References to undefined globals in function: tmpvar,tmpfn",
  fixed = TRUE
)

with_ref = .autocrate(
  function(y) {
    if (tmpvar == 10) tmpfn(y)
  },
  tmpvar = tmpvar,
  tmpfn=tmpfn
)

testthat::expect_equal(
  ls(rlang::fn_env(with_ref)),
  c("tmpfn", "tmpvar")
)

library(dplyr)
# Global definitions in dplyr must be explicitly assigned like in .
a = .autocrate(function(x) {
  Species = NULL
  x %>% pull(Species) %>% head(5)
})

testthat::expect_equal(a(iris), structure(
  c(1L, 1L, 1L, 1L, 1L),
  levels = c("setosa", "versicolor", "virginica"),
  class = "factor"
))

# Autocrating is recursive on functions supplied as globals:
not_crated_fn = function(x) {rnorm(x); return("success")}

crated_fn = .autocrate(function(x) {
   generator(x)
}, generator = not_crated_fn)

tmp = format(body(rlang::fn_env(crated_fn)$generator))

testthat::expect_equal(
  tmp,
  c("{", "    stats::rnorm(x)", "    return(\"success\")", "}")
)


}\if{html}{\out{</div>}}
}

