% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-abc.R
\name{.calculate_weights_smc}
\alias{.calculate_weights_smc}
\title{Calculate weights for particles in a new wave}
\usage{
.calculate_weights_smc(sim_df, epsilon, prev_sim_df = NULL, kernel_t)
}
\arguments{
\item{sim_df}{the output of a wave of simulation including a
\code{abc_summary_distance} column}

\item{epsilon}{epsilon is a tolerance threshold that controls how closely
simulated summaries must match the observed ones to be considered
plausible. This is in the unit of \code{abc_summary_distance}. Initially the 0.5
quantile of distances, in subsequent waves this might be decreased}

\item{prev_sim_df}{the output of a previous ABC wave including a
\code{abc_weight} column}

\item{kernel_t}{A kernel bandwidth parameter for proposals. This controls the
amount of noise that particles are perturbed by (and hence the spread of
the PDF of the proposal distribution), and 1 is approximately 34\% of the
proposal distribution at the centre. Smaller values (default is 0.2) give a
smaller step size in generating new proposals, and proposals will be closer
to currently accepted particles.}
}
\value{
the \code{sim_df} with an \code{abc_weight} column
}
\description{
The ABC weights need to be calculated for sampling from the proposal
distribution. They depend on the priors, an acceptance tolerance and the
weighted particles from the previous wave, which define the proposal
probability distribution.
}
\keyword{internal}
