% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-abc.R
\name{.compare_waves}
\alias{.compare_waves}
\title{Generate comparison metrics for two sequential waves}
\usage{
.compare_waves(sim_df, prev_sim_df = NULL, priors_list)
}
\arguments{
\item{sim_df}{the output of a wave of simulation including a
\code{abc_summary_distance} column}

\item{prev_sim_df}{the output of a previous ABC wave including a
\code{abc_weight} column}

\item{priors_list}{the list of priors as a named list of \code{dist_fn}s, plus one
sided formulae as constraints and 2 sided formulae as derived values.
A correlation matrix may also be present as an attribute \code{cor}.}
}
\value{
a nested tibble with 2 columns \code{summary} and \code{per_parameter}
with stats in each. The \code{summary} stats are
}
\description{
Generate comparison metrics for two sequential waves
}
\keyword{internal}
