% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-crate.R
\name{.crate_fn}
\alias{.crate_fn}
\title{Crate a pre-existing function with or without checks}
\usage{
.crate_fn(
  .fn,
  ...,
  .parent_env = baseenv(),
  .error_arg = ".fn",
  .error_call = environment(),
  .nsqualify = FALSE
)
}
\arguments{
\item{.fn}{a reference to a function}

\item{...}{Named arguments to declare in the environment of \code{.fn}.}

\item{.parent_env}{The default of \code{baseenv()} ensures that the evaluation
environment of the crate is isolated from the search path. Specifying
another environment such as the global environment allows this condition to
be relaxed (but at the expense of no longer being able to rely on a local
run giving the same results as one in a different process).}

\item{.error_arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{.error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{.nsqualify}{check for undefined global references and qualify namespaces?}
}
\value{
a crated function
}
\description{
Crate a pre-existing function with or without checks
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

.crate_fn
}\if{html}{\out{</div>}}
}

