% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-abc.R
\name{.distance_covariance}
\alias{.distance_covariance}
\title{Covariance from component scores}
\usage{
.distance_covariance(component_scores)
}
\arguments{
\item{component_scores}{a list of lists of scores, one per simulation}
}
\value{
a covariance matrix
}
\description{
Supports calculation of Mahalanobis distance.
}
\keyword{internal}
