% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-crate.R
\name{.DollarNames.crate}
\alias{.DollarNames.crate}
\title{Support for auto suggests on \code{crate}s}
\usage{
\method{.DollarNames}{crate}(x, pattern)
}
\arguments{
\item{x}{a crated function}

\item{pattern}{a regular expression}
}
\value{
the names of crated data items
}
\description{
Support for auto suggests on \code{crate}s
}
\concept{crate_utils}
\keyword{internal}
