% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-ggplot-utils.R
\name{.gg_hist}
\alias{.gg_hist}
\alias{.gg_dens}
\alias{.gg_bar}
\alias{.gg_scatter}
\alias{.gg_denshist}
\alias{.gg_linebar}
\alias{.gg_fun}
\title{Simple one liner plots from vectors with defaults}
\usage{
.gg_hist(x, binwidth = 1, ...)

.gg_dens(x, bw = NULL, ...)

.gg_bar(x, ...)

.gg_scatter(x, y, ...)

.gg_denshist(x, y, binwidth = 1)

.gg_linebar(x, y, binwidth = 1)

.gg_fun(fn, ..., xlim = NULL)
}
\arguments{
\item{x}{a vector of data or counts}

\item{binwidth}{histogram binwidth}

\item{...}{passed onto the geoms}

\item{bw}{bandwidth parameter,if given it is interpreted as a fraction of the sd of the data}

\item{y}{a vector or matrix of data or counts}
}
\value{
a ggplot
}
\description{
Simple one liner plots from vectors with defaults
}
\section{Functions}{
\itemize{
\item \code{.gg_hist()}: Simple histogram

\item \code{.gg_dens()}: A density plot

\item \code{.gg_bar()}: A bar chart of counts

\item \code{.gg_scatter()}: X-Y scatter plot

\item \code{.gg_denshist()}: Histogram with densities overlaid.

\item \code{.gg_linebar()}: Bar chart with lines overlaid

\item \code{.gg_fun()}: Function plot

}}
\concept{ggplot}
\keyword{internal}
