% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-ggplot-utils.R
\name{.gg_merge_aes}
\alias{.gg_merge_aes}
\title{Merge aesthetic mappings and deduplicate the result}
\usage{
.gg_merge_aes(..., allow_override = TRUE)
}
\arguments{
\item{...}{a set of \code{name=value}, and \code{aes(...)} specifications}

\item{allow_override}{are the original mappings kept and \code{name=value} pairs
act as default and may be overridden by user supplied aesthetics}
}
\value{
a single deduplicated set.
}
\description{
Merge aesthetic mappings and deduplicate the result
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

m1 = ggplot2::aes(x=a,y=b,colour=class)
.gg_merge_aes(x=A,y=BB,m1)
}\if{html}{\out{</div>}}
}

