% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-empirical.R
\name{.interpolate}
\alias{.interpolate}
\title{Interpolate from 2 coordinates and predict output}
\usage{
.interpolate(y, x, new_x = NULL)
}
\arguments{
\item{y}{the y values. length 2.}

\item{x}{the x values. length 2.}

\item{new_x}{the values to predict at (vectorised). If null will return a
function that can predict new values.}
}
\value{
the interpolated value or a prediction function
}
\description{
Interpolate from 2 coordinates and predict output
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

.interpolate(c(0,1),c(0,1),5)
}\if{html}{\out{</div>}}
}

