% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{.log_kernel}
\alias{.log_kernel}
\title{Generate a non normalised log of kernel for ABC weight calculation}
\usage{
.log_kernel(
  u,
  epsilon,
  kernel = c(c("epanechnikov", "uniform", "triangular", "biweight", "gaussian"))
)
}
\arguments{
\item{u}{the distance (dissimilarilty) between}

\item{epsilon}{epsilon is a tolerance threshold that controls how closely
simulated summaries must match the observed ones to be considered
plausible. This is in the unit of \code{abc_summary_distance}. Initially the 0.5
quantile of distances, in subsequent waves this might be decreased.
It is the scale parameter of the kernel function. $K_h(|u|)$}

\item{kernel}{one of "epanechnikov", "uniform", "triangular", "biweight",
"gaussian"}
}
\value{
the log of the kernel function
}
\description{
\deqn{
\alpha_{i,j} = \frac{1}{\beta+1}\\
\beta = \alpha +1
}
}
\keyword{internal}
