% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-recycle.R
\name{.make_square}
\alias{.make_square}
\title{Strictly recycle the items in a list to the same length}
\usage{
.make_square(lst)
}
\arguments{
\item{lst}{a list input}
}
\value{
the list with all items recycled to the same length
}
\description{
This removes NULL entries, and zero length lists
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

lst = list(a = 1,b=1:3, c= NULL, d=list(), e=character())
testthat::expect_equal(
  .make_square(lst),
  list(a = c(1, 1, 1), b = 1:3)
)
}\if{html}{\out{</div>}}
}

