% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-empirical.R
\name{.monotonicpolyspline}
\alias{.monotonicpolyspline}
\title{Create a \code{polySpline} object from a \code{stats::splinefun} call.}
\usage{
.monotonicpolyspline(x, y, ties = mean)
}
\arguments{
\item{x, y}{vectors giving the coordinates of the points to be
    interpolated.  Alternatively a single plotting structure can be
    specified: see \code{\link{xy.coords}}.

    \code{y} must be increasing or decreasing for \code{method = "hyman"}.
  }

\item{ties}{handling of tied \code{x} values.  The string
    \code{"ordered"} or a function (or the name of a function) taking a
    single vector argument and returning a single number or a length-2
    \code{\link{list}} of both, see \code{\link[stats]{approx}} and its
    \sQuote{Details} section, and the example below.}
}
\value{
An object of class \code{polySpline}.
}
\description{
Create a \code{polySpline} object from a \code{stats::splinefun} call.
}
\details{
This function converts the output of \code{stats::splinefun}, to a
a polynomial spline object from the \code{splines} package.
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{


# strictly increasing
spl = .monotonicpolyspline(1:10, log(1:10))
testthat::expect_equal(
  predict(spl,4.5)$y,
  log(4.5),
  tolerance = 0.01
)

# strictly decreasing
spl2 = .monotonicpolyspline(1:10, -log(1:10))
testthat::expect_equal(
  predict(spl2,4.5)$y,
  -log(4.5),
  tolerance = 0.01
)

# not monotonic
testthat::expect_error(
  {
    .monotonicpolyspline(-4:4, (-4:4)^2)
  },
  "Data is not monotonic.",
  fixed = TRUE
)

}\if{html}{\out{</div>}}
}

