% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-sigmoid.R
\name{.p_from_n}
\alias{.p_from_n}
\title{Proportions from a vector of lengths}
\usage{
.p_from_n(n, n_50, n_01, p_max = 1, n_100 = 0)
}
\arguments{
\item{n}{a vector of lengths}

\item{n_50}{the size at 50 percent}

\item{n_01}{the size at 1 percent, and the limit of the returned \code{p} value
such that \eqn{p \times n \ge n_{01}}}

\item{p_max}{the maximum value at \code{n <= n_100}}

\item{n_100}{an offset under which the value is 1 (default is 0)}
}
\value{
a vector of proportions for each length
}
\description{
Typically useful in finding alpha values for plotting, or bandwidths for
smoothing data this is a decreasing sigmoid with tunable parameters.
}
\details{
This is using a composite of a
\href{https://en.wikipedia.org/wiki/Hill_equation_(biochemistry)}{Hill function}
and a reciprocal:

\deqn{
f(n) = \frac{1}{1 + \left( \dfrac{n - n_{100}}{n_{50}} \right)^b} \times p_{max}\\
b = \frac{\ln(99)}{\ln(n_{01}/n_{50})}
}

If \eqn{n-n_{100} \le 0} then the maximum value is returned
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{


testthat::expect_equal(.p_from_n(c(0, 5, 100), 5, 100), c(1, 0.5, 0.01))

testthat::expect_error(
  {
    .p_from_n(0, 50, 1)
  },
  "inconsistent parameters, n_01 must be larger than n_50.",
  fixed = TRUE
)

testthat::expect_equal(.p_from_n(c(0:10), 5, 10, 0.1), c(
  0.1,
  0.0999976758268449,
  0.0997704297474442,
  0.0967278219151741,
  0.081446655140766,
  0.05,
  0.0229935645768342,
  0.0097036542590398,
  0.00424593238101868,
  0.00199056073397113,
  0.001
))

# does not go below
testthat::expect_equal(
  .p_from_n(c(1000, 10000, 100000), 5, 50),
  c(5e-04, 5e-05, 5e-06)
)

}\if{html}{\out{</div>}}
}

