% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-sigmoid.R
\name{.p_from_n2}
\alias{.p_from_n2}
\title{Simple sigmoid tending towards a minimum proportion of input}
\usage{
.p_from_n2(n, n_inf, p_max = 1, n_100 = 0)
}
\arguments{
\item{n}{a vector of lengths}

\item{p_max}{the maximum value at \code{n <= n_100}}

\item{n_100}{an offset under which the value is 1 (default is 0)}
}
\value{
a vector of proportions for each length
}
\description{
Simple sigmoid tending towards a minimum proportion of input
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

testthat::expect_equal(.p_from_n2(0:5, 1), c(
  1,
  0.707106781186547,
  0.447213595499958,
  0.316227766016838,
  0.242535625036333,
  0.196116135138184
))
}\if{html}{\out{</div>}}
}

