% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-abc.R
\name{.posterior_covariance}
\alias{.posterior_covariance}
\title{Covariance matrix from the posteriors}
\usage{
.posterior_covariance(theta, weights, kernel_t = 1)
}
\arguments{
\item{theta}{a set of particles as a matrix}

\item{weights}{the weights of the previous particles assumed calculated using
the same kernel}

\item{kernel_t}{A kernel bandwidth parameter for proposals. This controls the
amount of noise that particles are perturbed by (and hence the spread of
the PDF of the proposal distribution), and 1 is approximately 34\% of the
proposal distribution at the centre. Smaller values (default is 0.2) give a
smaller step size in generating new proposals, and proposals will be closer
to currently accepted particles.}
}
\value{
a covariance matrix, optionally scaled
}
\description{
Covariance matrix from the posteriors
}
\keyword{internal}
