% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-empirical.R
\name{.predict_lm_1d}
\alias{.predict_lm_1d}
\title{Fit a 1D linear model and predict output}
\usage{
.predict_lm_1d(y, x, new_x = NULL)
}
\arguments{
\item{y}{the y values. At least 2.}

\item{x}{the x values. At least 2.}

\item{new_x}{the values to predict at (vectorised). If null will return a
function that can predict new values.}
}
\value{
the predicted value or a prediction function
}
\description{
Fit a 1D linear model and predict output
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

.predict_lm_1d(c(6,5,7,10), c(1,2,3,4), c(0,1))
.predict_lm_1d(c(0,1), c(0,1), -2:2)
}\if{html}{\out{</div>}}
}

