% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-recycle.R
\name{.recycle_and_name}
\alias{.recycle_and_name}
\title{Recycles an input to match the length a set of names}
\usage{
.recycle_and_name(input, names)
}
\arguments{
\item{input}{an input vector which may be named}

\item{names}{a set of names to match}
}
\value{
a recycled vector with names matching the order of \code{names}
}
\description{
Recycles an input to match the length a set of names
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{


testthat::expect_equal(
  .recycle_and_name(c(a = 1, b = 2, c = 3), c("c", "b", "a")),
  c(c = 3, b = 2, a = 1)
)

testthat::expect_equal(
  .recycle_and_name(1, c("A", "B")),
  c(A = 1, B = 1)
)

}\if{html}{\out{</div>}}
}

