% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-abc_prior-class.R
\name{.sample_priors}
\alias{.sample_priors}
\title{Sample for the prior distribution}
\usage{
.sample_priors(proposal_list, n_sims)
}
\arguments{
\item{proposal_list}{a list of empirical probability distributions that map
MVN space to proposal space, and are the "prior" for each adaptive wave.
This is already used to generate the proposals and their mapping in \code{sim_df}}

\item{n_sims}{The number of simulations to run per wave (for SMC and Adaptive)
or overall (for Rejection). For rejection sampling a large number is
recommended, for the others sma}
}
\value{
a data frame of samples in MVN (prefixed \code{abc_mvn_}) and parameter
space.
}
\description{
This uses a multivariate normal and a copula to generate correlated
structure. The correlation is held as an attribute in the proposals.
N.B. this used to construct derived values and apply constraints but this
was complex and has been deferred. Constraints change the probability
distribution also and will affect validity of proposal distribution.
}
\concept{abc_prior_s3}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

p = new_abc_prior(
  .dists = list(
    mean = as.dist_fns("norm",4,2),
    sd = as.dist_fns("gamma",2)
  ),
  .derived = list(
    shape ~ mean^2 / sd^2,
    rate ~ mean / sd^2
  )
)

s = .sample_priors(p,10)
testthat::expect_equal(
  colnames(s),
  c("abc_mvn_mean", "mean", "abc_mvn_sd", "sd")
)
}\if{html}{\out{</div>}}
}

