% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-abc.R
\name{.summary_distance}
\alias{.summary_distance}
\title{Combine simulation scores to a single distance metric.}
\usage{
.summary_distance(
  component_scores,
  obsscores = NULL,
  distance_method = c("euclidean", "manhattan", "mahalanobis"),
  wave1_cov = NULL,
  scoreweights = NULL
)
}
\arguments{
\item{component_scores}{\itemize{
\item a list column of scores. each entry is itself a list
}}

\item{obsscores}{Summary scores for the observational data. This will
be a named list, and is equivalent to the output of \code{scorer_fn},
on the observed data. If not given typically it will be assumed to be all
zeros.}

\item{distance_method}{what metric is used to combine \code{simscores} and \code{obsscores}
and is one of \code{"euclidean"}, \code{"manhattan"}, or \code{"mahalanobis"}.}

\item{scoreweights}{A named vector with names matching output of \code{scorer_fn}
that defines the importance of this component of the scoring in the overall
distance and weighting of any given simulation. This can be used to assign
more weight on certain parts of the model output.}
}
\value{
a vector of distances
}
\description{
This function calculates the overall score for an individual simulation,
based on combining the components output by \code{score_list} comparing them to
the target component scores of the original observed data. The Mahalanobis
distance is calculated relative to the distribution of the first wave (i.e.
when the priors were directly sampled).
}
\keyword{internal}
