% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-crate.R
\name{.super_crate}
\alias{.super_crate}
\title{Create an environment of shared crates.}
\usage{
.super_crate(
  .fns,
  ...,
  .parent_env = baseenv(),
  .error_arg = ".fn",
  .error_call = environment(),
  .nsqualify = FALSE
)
}
\arguments{
\item{.fns}{a named list of functions}

\item{...}{Named arguments to declare in the environment of \code{.fn}.}

\item{.parent_env}{The default of \code{baseenv()} ensures that the evaluation
environment of the crate is isolated from the search path. Specifying
another environment such as the global environment allows this condition to
be relaxed (but at the expense of no longer being able to rely on a local
run giving the same results as one in a different process).}

\item{.error_arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{.error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{.nsqualify}{check for undefined global references and qualify namespaces?}
}
\value{
an environment containing crated functions
}
\description{
Crated functions that share the same data in a single dependency free
environment.
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

msg = "hello"
e = .super_crate(
  list(
    plusx = function(z) {
     return(z+x)
    },
    plusy = function(z) {
     return(z+y+plusx(z))
    }
    # splicing does not work
    # hi = function() {
    #   print(!!msg)
    # }
  ),
  x = 1:10,
  y = 11:20
)

testthat::expect_equal(
  e$plusy(1),
  c(14, 16, 18, 20, 22, 24, 26, 28, 30, 32)
)
}\if{html}{\out{</div>}}
}

