% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-empirical.R
\name{empirical_data}
\alias{empirical_data}
\title{Fit a piecewise logit transformed linear model to weighted data}
\usage{
empirical_data(x, w = NULL, link = "identity", ..., name = NULL, bw = NULL)
}
\arguments{
\item{x}{either a vector of samples from a distribution \code{X} or cut-offs for
cumulative probabilities when combined with \code{p}}

\item{w}{for data fits, a vector the same length as \code{x} giving the
importance weight of each observation. This does not need to be normalised.
There must be some non zero weights, and all must be finite.}

\item{link}{a link function. Either as name, or a \code{link_fns} S3 object. In
the latter case this could be derived from a statistical distribution by
\verb{as.link_fns(<dist_fns>)}. This supports the use of a prior to define the
support of the empirical function, and is designed to prevent tail
truncation. Support for the updated quantile function will be the same as
the provided prior.}

\item{...}{
  Named arguments passed on to \code{\link[=.logit_z_interpolation]{.logit_z_interpolation}}\describe{
    \item{\code{bw}}{a bandwidth expressed in terms of the probability width, or proportion
of observations.}
}
  Named arguments passed on to \code{\link[=empirical_cdf]{empirical_cdf}}\describe{
    \item{\code{smooth}}{fits the empirical distribution with a pair of splines for CDF
and quantile function, creating a mostly smooth density. This smoothness
comes at the price of potential over-fitting and will produce small
differences between \code{p} and \code{q} functions such that \code{x=p(q(x))} is no
longer exactly true. Setting this to false will replace this with a
piecewise linear fit that is not smooth in the density, but is exact in
forward and reverse transformation.}
\item{\code{...}}{not used}
}}

\item{name}{a label for this distribution. If not given one will be generated
from the distribution and parameters. This can be used as part of plotting.}

\item{bw}{a bandwidth expressed in terms of the probability width, or proportion
of observations.}
}
\value{
a \code{dist_fns} S3 object that function that contains statistical
distribution functions for this data.
}
\description{
This fits a CDF and quantile function to ranked data in a
transformed space. X value transformation is specified in the \code{link}
parameter and is either something like "log", "logit", etc.
}
\details{
The empirical distribution fitted is a piecewise linear in z transformed X
and logit Y space. The evaluation points are linearly interpolated in this
space given a bandwidth for interpolation.

This function imputes tails of distributions. Given perfect data as samples
or as quantiles it should recover the tail
}
\concept{empirical}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{


# from samples:
withr::with_seed(123,{
 e2 = empirical_data(rnorm(10000), bw=0.1)
 testthat::expect_equal(e2$p(-5:5), pnorm(-5:5), tolerance=0.01)
 testthat::expect_equal(e2$d(-5:5), dnorm(-5:5), tolerance=0.05)
 testthat::expect_equal(e2$q(seq(0,1,0.1)), qnorm(seq(0,1,0.1)), tolerance=0.025)
})


# Construct a normal using a sequence and density as weight.
e7 = empirical_data(
  x=seq(-10,10,length.out=1000),
  w=dnorm(seq(-10,10,length.out=1000))
)
testthat::expect_equal(e7$p(-5:5), pnorm(-5:5), tolerance=0.01)

}\if{html}{\out{</div>}}
}

