% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa-examples.R
\name{example_fns}
\alias{example_fns}
\alias{example_sim_fn}
\alias{example_scorer_fn}
\alias{example_obs}
\alias{example_truth}
\alias{example_obsdata}
\alias{example_priors_list}
\alias{example_rejection_fit}
\alias{example_smc_fit}
\alias{example_adaptive_fit}
\title{Example generators}
\usage{
example_sim_fn(mean, sd1, sd2)

example_scorer_fn(obsdata, simdata)

example_obs()

example_truth()

example_obsdata()

example_priors_list()

example_rejection_fit()

example_smc_fit()

example_adaptive_fit()
}
\arguments{
\item{mean}{the example mean}

\item{sd1}{the example normal sd (A)}

\item{sd2}{the example gamma sd (B)}

\item{obsdata}{e.g. example_obsdata()}

\item{simdata}{output of example_sim_fn()}
}
\value{
a list of A and B of samples from a normal and gamma

a list of scores

output of \code{test_simulation()}

example \code{truth} parameter

example \code{obsdata} parameter

example \code{priors_list} parameter

example \code{abc_rejection} output

example \code{abc_smc} output

example \code{abc_adaptive} output
}
\description{
These are a set of internally cached functions to support
examples. They are exported as internal functions so that the examples can
run correctly and cache their output to prevent excessive repetition of the
code examples.
}
\section{Functions}{
\itemize{
\item \code{example_sim_fn()}: Example simulation function

\item \code{example_scorer_fn()}: Example simulation function

\item \code{example_obs()}: Example output of \code{test_simulation()}

\item \code{example_truth()}: Example for \code{truth} parameter

\item \code{example_obsdata()}: Example for \code{obsdata} parameter

\item \code{example_priors_list()}: Example for \code{priors_list} parameter

\item \code{example_rejection_fit()}: Example \code{abc_rejection} output

\item \code{example_smc_fit()}: Example \code{abc_smc} output

\item \code{example_adaptive_fit()}: Example \code{abc_adaptive} output

}}
\examples{
example_sim_fn(3,2,1) \%>\% lapply(head,10)
example_scorer_fn(
  example_obsdata(),
  example_sim_fn(3,2,1)
)
example_obs() \%>\% rapply(head,n=10, how="replace")
example_truth()
example_obsdata() \%>\% lapply(head,10)
example_priors_list()
example_rejection_fit()
example_smc_fit()
example_adaptive_fit()
}
\keyword{internal}
