% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-workflow.R
\name{fixed_wave_termination_fn}
\alias{fixed_wave_termination_fn}
\title{Run the SMC or adaptive algorithm for a set number of waves}
\usage{
fixed_wave_termination_fn(max_wave)
}
\arguments{
\item{max_wave}{the number of waves to run}
}
\value{
A function that will test for completion
}
\description{
Run the SMC or adaptive algorithm for a set number of waves
}
\examples{
# Declare converged after 3 waves:
converged_fn = fixed_wave_termination_fn(3)
}
