% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{format.dist_fns}
\alias{format.dist_fns}
\title{Format a \code{dist_fns} S3 object}
\usage{
\method{format}{dist_fns}(x, ..., digits = 3)
}
\arguments{
\item{x}{a \code{dist_fns} S3 object}

\item{...}{passed onto methods}

\item{digits}{the number of significant digits}
}
\value{
a character value
}
\description{
Format a \code{dist_fns} S3 object
}
\concept{dist_fns_s3}
