% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{format.link_fns}
\alias{format.link_fns}
\title{Format a \code{link_fns} S3 object}
\usage{
\method{format}{link_fns}(x, ...)
}
\arguments{
\item{x}{a \code{link_fns} S3 object}

\item{...}{passed onto methods}
}
\value{
a character value
}
\description{
Format a \code{link_fns} S3 object
}
\concept{link_fns_s3}
