% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{is.dist_fns_list}
\alias{is.dist_fns_list}
\title{Check if this is a \code{dist_fns_list} S3 object}
\usage{
is.dist_fns_list(x)
}
\arguments{
\item{x}{anything}
}
\value{
TRUE or FALSE
}
\description{
Check if this is a \code{dist_fns_list} S3 object
}
\concept{dist_fns_s3}
