% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{is.link_fns}
\alias{is.link_fns}
\title{Check if this is a \code{link_fns} S3 object}
\usage{
is.link_fns(x)
}
\arguments{
\item{x}{anything}
}
\value{
TRUE or FALSE
}
\description{
Check if this is a \code{link_fns} S3 object
}
\concept{link_fns_s3}
