% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-empirical.R
\name{kurtosis}
\alias{kurtosis}
\title{Calculate the excess kurtosis of a set of data}
\usage{
kurtosis(x, na.rm = FALSE, excess = TRUE)
}
\arguments{
\item{x}{a vector of observations}

\item{na.rm}{remove \code{NA}s?}

\item{excess}{if false calculates raw kurtosis rather than excess}
}
\value{
the excess kurtosis
}
\description{
Calculate the excess kurtosis of a set of data
}
\examples{
kurtosis(rnorm(1000))
kurtosis(rpois(1000, 2)) # leptokurtic > 0 (usually)
kurtosis(runif(1000)) # platykurtic: < 0

kurtosis(rlnorm(1000))
}
\concept{empirical}
