% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{length.dist_fns}
\alias{length.dist_fns}
\title{\code{dist_fns} S3 objects length is always 1}
\usage{
\method{length}{dist_fns}(x, ...)
}
\arguments{
\item{x}{a \code{dist_fns} S3 object}

\item{...}{passed onto methods}
}
\value{
1 always
}
\description{
Boilerplate S3 class operation
}
\concept{dist_fns_s3}
\keyword{internal}
