% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{length.link_fns}
\alias{length.link_fns}
\title{\code{link_fns} S3 objects length is always 1}
\usage{
\method{length}{link_fns}(x, ...)
}
\arguments{
\item{x}{a \code{link_fns} S3 object}

\item{...}{passed onto methods}
}
\value{
1 always
}
\description{
Boilerplate S3 class operation
}
\concept{link_fns_s3}
\keyword{internal}
