% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{length.link_fns_list}
\alias{length.link_fns_list}
\title{Length of a \code{link_fns_list}}
\usage{
\method{length}{link_fns_list}(x, ...)
}
\arguments{
\item{x}{a \code{link_fns_list} S3 object}

\item{...}{passed onto methods}
}
\value{
the length
}
\description{
These boilerplate functions allow generic list behaviour from \code{link_fns_list}
classes allowing \code{link_fns} S3 objects to be used in lists or dataframes.
}
\concept{link_fns_s3}
\keyword{internal}
