% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{link_fns}
\alias{link_fns}
\title{Create an empty \code{link_fns_list}}
\usage{
link_fns()
}
\value{
an empty \code{link_fns_list}
}
\description{
Boilerplate S3 class operation
}
\concept{link_fns_s3}
