% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{logit-norm}
\alias{logit-norm}
\title{Logit-normal distribution}
\arguments{
\item{n}{number of observations}

\item{x}{vector of quantiles \verb{(0<x<1)}}

\item{q}{vector of quantiles \verb{(0<q<1)}}

\item{p}{vector of probabilities}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X<=x]}
otherwise \code{P[X>x]}.}

\item{meanlogit}{the mean on the logit scale}

\item{sdlogit}{the sd on the logit scale}

\item{prob.0.5}{the median on the true scale}

\item{kappa}{a dispersion parameter from 0 (none) to 1 maximum dispersion}
}
\value{
a vector of probabilities, quantiles, densities or samples.
}
\description{
The logit-normal distribution has a support of 0 to 1.
}
\concept{distributions}
\keyword{internal}
