% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{map_link_fns}
\alias{map_link_fns}
\title{Apply a function to each element of a vector returning a \code{link_fns_list}}
\usage{
map_link_fns(.x, .f, ..., .progress = FALSE)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.f}{a function to apply that returns a \code{link_fns} S3 object (usually an
\code{as.link_fns()} call)}

\item{...}{Additional arguments passed on to the mapped function.

We now generally recommend against using \code{...} to pass additional
(constant) arguments to \code{.f}. Instead use a shorthand anonymous function:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{# Instead of
x |> map(f, 1, 2, collapse = ",")
# do:
x |> map(\\(x) f(x, 1, 2, collapse = ","))
}\if{html}{\out{</div>}}

This makes it easier to understand which arguments belong to which
function and will tend to yield better error messages.}

\item{.progress}{Whether to show a progress bar. Use \code{TRUE} to turn on
a basic progress bar, use a string to give it a name, or see
\link[purrr]{progress_bars} for more details.}
}
\value{
a \code{link_fns_list}
}
\description{
Analogous to \code{purrr::map_dbl()}
}
\seealso{
\code{\link[purrr:map]{purrr::map()}}
}
\concept{link_fns_s3}
