% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-mixture.R
\name{mixture}
\alias{mixture}
\title{Construct a mixture distribution}
\usage{
mixture(dists, weights = 1, steps = 200, tail_p = 1e-05)
}
\arguments{
\item{dists}{a \code{dist_fn_list} of distribution functions}

\item{weights}{a vector of weights}
}
\value{
a \code{dist_fn} of the mixture distribution
}
\description{
Construct a mixture distribution
}
\concept{empirical}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{


dists = c(
  as.dist_fns("norm", mean=-1),
  as.dist_fns("norm", mean=1),
  as.dist_fns("gamma", shape=2)
)
weights = c(1,1,0.3)

mix = mixture(dists,weights)
testthat::expect_equal(
  format(mix),
  "mixture; Median (IQR) 0.289 [-0.886 — 1.31]"
)

}\if{html}{\out{</div>}}
}

