% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-abc_fit-class.R
\name{new_abc_fit}
\alias{new_abc_fit}
\title{Create a \code{abc_fit} object}
\usage{
new_abc_fit(type, iterations, converged, priors_list, wave_df, summ_df, sim_df)
}
\arguments{
\item{type}{the type of ABC algorithm}

\item{iterations}{number of completed iterations}

\item{converged}{boolean - did the result meet convergence criteria}

\item{wave_df}{a list of dataframes of wave convergence metrics}

\item{summ_df}{the summary of the parameter fits after each wave.}

\item{sim_df}{the final wave posteriors}
}
\value{
an S3 ojbect of class \code{abc_fit} this contains the following:
\itemize{
\item type: the type of ABC algorithm
\item iterations: number of completed iterations
\item converged: boolean - did the result meet convergence criteria
\item waves: a list of dataframes of wave convergence metrics
\item summary: a dataframe with the summary of the parameter fits after each wave.
\item posteriors: the final wave posteriors
}
}
\description{
A class holding the output of a single ABC model fitting, either as a
single ABC rejection round or after a set of SMC waves.
}
\concept{abc_fit_s3}
\keyword{internal}
