% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{new_dist_calls}
\alias{new_dist_calls}
\title{Create a new \code{dist_fns} S3 object}
\usage{
new_dist_calls(
  name = "NULL",
  pcall,
  qcall,
  rcall,
  dcall,
  ...,
  params = NULL,
  knots = NULL
)
}
\arguments{
\item{name}{a name for the distribution}

\item{pcall}{a call of the cumulative probability function}

\item{qcall}{a call of the quantile function}

\item{rcall}{a call of the RNG function}

\item{dcall}{a call of the density function}

\item{...}{parameters for the distribution.}

\item{params}{parameters for the distribution as a named list.}

\item{knots}{in empirical distributions this holds details on the knot points}
}
\value{
a new \code{dist_fns} S3 object or a \code{dist_fns_list} if parameters are vectorised
}
\description{
Create a new \code{dist_fns} S3 object
}
\concept{dist_fns_s3}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

tmp = new_dist_calls(name="norm",
  pcall=.qual("pnorm"),
  qcall=.qual("qnorm"),
  rcall=.qual("rnorm"),
  dcall=.qual("dnorm")
)
testthat::expect_equal(tmp$p(-2:2), c(
  0.0227501319481792,
  0.158655253931457,
  0.5,
  0.841344746068543,
  0.977249868051821
))
}\if{html}{\out{</div>}}
}

