% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{new_dist_fns}
\alias{new_dist_fns}
\title{Create a new \code{dist_fns} S3 object}
\usage{
new_dist_fns(
  name = "NULL",
  pfn = pnull,
  qfn = qnull,
  rfn = NULL,
  dfn = NULL,
  ...,
  params = NULL,
  knots = NULL
)
}
\arguments{
\item{name}{a name for the distribution}

\item{pfn}{the cumulative probability function}

\item{qfn}{the quantile function}

\item{rfn}{the RNG function}

\item{dfn}{the density function}

\item{...}{parameters for the distribution.}

\item{params}{parameters for the distribution as a named list.}

\item{knots}{in empirical distributions this holds details on the knot points}
}
\value{
a new \code{dist_fns} S3 object or a \code{dist_fns_list} if parameters are vectorised
}
\description{
Create a new \code{dist_fns} S3 object
}
\concept{dist_fns_s3}
\keyword{internal}
