% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{pcgamma}
\alias{pcgamma}
\title{Cumulative probability: gamma distribution constrained to have mean > sd}
\usage{
pcgamma(q, mean, kappa = 1/mean, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantiles}

\item{mean}{the mean value on the true scale (vectorised)}

\item{kappa}{a coefficient of variation. where 0 is no variability and 1 is
maximally variability (vectorised)}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X<=x]}
otherwise \code{P[X>x]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
\code{dgamma} gives the density,
  \code{pgamma} gives the distribution function,
  \code{qgamma} gives the quantile function, and
  \code{rgamma} generates random deviates.

  Invalid arguments will result in return value \code{NaN}, with a warning.

  The length of the result is determined by \code{n} for
  \code{rgamma}, and is the maximum of the lengths of the
  numerical arguments for the other functions.

  The numerical arguments other than \code{n} are recycled to the
  length of the result.  Only the first elements of the logical
  arguments are used.
}
\description{
Density, distribution function, quantile function and random
  generation for the Gamma distribution with parameters \code{shape} and
  \code{scale}.
}
\examples{
pcgamma(seq(0,4,0.25), 2, 0.5)
}
\seealso{
\code{\link[stats:GammaDist]{stats::pgamma()}}
}
\concept{distributions}
