% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{pillar_shaft.dist_fns_list}
\alias{pillar_shaft.dist_fns_list}
\title{Pillar shaft function}
\usage{
\method{pillar_shaft}{dist_fns_list}(x, ...)
}
\arguments{
\item{x}{a \code{dist_fns_list} S3 object}

\item{...}{passed onto methods}
}
\value{
a pillar shaft
}
\description{
These boilerplate functions allow generic list behaviour from \code{dist_fns_list}
classes allowing \code{dist_fns} S3 objects to be used in lists or dataframes.
}
\concept{dist_fns_s3}
\keyword{internal}
