% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{plogitnorm}
\alias{plogitnorm}
\title{Logit-normal distribution}
\usage{
plogitnorm(q, meanlogit = 0, sdlogit = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantiles \verb{(0<q<1)}}

\item{meanlogit}{the mean on the logit scale}

\item{sdlogit}{the sd on the logit scale}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X<=x]}
otherwise \code{P[X>x]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
a vector of probabilities, quantiles, densities or samples.
}
\description{
The logit-normal distribution has a support of 0 to 1.
}
\examples{
plogitnorm(seq(0.1,0.9,0.1), 0, 1)
}
\concept{distributions}
