% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{plogitnorm2}
\alias{plogitnorm2}
\title{Logit-normal distribution}
\usage{
plogitnorm2(
  q,
  prob.0.5 = 0.5,
  kappa = 1 - exp(-1),
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{q}{vector of quantiles \verb{(0<q<1)}}

\item{prob.0.5}{the median on the true scale}

\item{kappa}{a dispersion parameter from 0 (none) to 1 maximum dispersion}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X<=x]}
otherwise \code{P[X>x]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
a vector of probabilities, quantiles, densities or samples.
}
\description{
The logit-normal distribution has a support of 0 to 1.
}
\examples{
plogitnorm2(seq(0.1,0.9,0.1), 0.75, 0.2)
plogitnorm2(qlogitnorm2(seq(0.1,0.9,0.1), 0.75, 0.2), 0.75, 0.2)
}
\concept{distributions}
