% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{plot.dist_fns}
\alias{plot.dist_fns}
\title{Plot a \code{dist_fns} S3 object}
\usage{
\method{plot}{dist_fns}(x, ...)
}
\arguments{
\item{x}{a \code{dist_fns} S3 object}

\item{...}{
  Named arguments passed on to \code{\link[=plot.dist_fns_list]{plot.dist_fns_list}}\describe{
    \item{\code{x}}{a \code{dist_fns_list}}
\item{\code{...}}{passed to \code{ggplot2::geom_step},\code{ggplot2::geom_rect} or \code{ggplot2::geom_area},}
\item{\code{mapping}}{override default aesthetics with \code{name}, \code{id} or \code{group}}
\item{\code{steps}}{resolution of the plot}
\item{\code{tail}}{the minimum tail probability to plot}
\item{\code{plot_quantiles}}{by default the quantiles of the distribution are
plotted over the density sometimes this makes it hard to read}
\item{\code{smooth}}{by default some additional smoothing is used to cover up small
discontinuities in the PDF.}
}}
}
\value{
a ggplot
}
\description{
Plot a \code{dist_fns} S3 object
}
\concept{dist_fns_s3}
