% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{plot.dist_fns_list}
\alias{plot.dist_fns_list}
\title{Plot a \code{dist_fns_list} S3 object}
\usage{
\method{plot}{dist_fns_list}(
  x,
  ...,
  mapping = .gg_check_for_aes(...),
  steps = 200,
  tail = 0.001,
  plot_quantiles = TRUE,
  smooth = TRUE
)
}
\arguments{
\item{x}{a \code{dist_fns_list}}

\item{...}{passed to \code{ggplot2::geom_step},\code{ggplot2::geom_rect} or \code{ggplot2::geom_area},}

\item{mapping}{override default aesthetics with \code{name}, \code{id} or \code{group}}

\item{steps}{resolution of the plot}

\item{tail}{the minimum tail probability to plot}

\item{plot_quantiles}{by default the quantiles of the distribution are
plotted over the density sometimes this makes it hard to read}

\item{smooth}{by default some additional smoothing is used to cover up small
discontinuities in the PDF.}
}
\value{
a ggplot
}
\description{
Plot a smoothed version of the PDFs of a set of \code{dist_fns}. These are \code{ggplots}
that can be facetted by \code{names}, \code{id} or \code{group}
}
\concept{dist_fns_s3}
