% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-convergence.R
\name{plot_convergence}
\alias{plot_convergence}
\title{Plot convergence metrics by wave for SMC and adaptive ABC}
\usage{
plot_convergence(fit)
}
\arguments{
\item{fit}{A S3 \code{abc_fit} object as output by the \code{abc_XXX} functions}
}
\value{
a patchwork plot of convergence metrics
}
\description{
Plot convergence metrics by wave for SMC and adaptive ABC
}
\examples{

plot_convergence(
  example_adaptive_fit()
)

}
\concept{workflow}
