% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-convergence.R
\name{plot_evolution}
\alias{plot_evolution}
\title{Plot the evolution of the density function by wave for SMC and adaptive ABC}
\usage{
plot_evolution(fit, truth = NULL, ...)
}
\arguments{
\item{fit}{A S3 \code{abc_fit} object as output by the \code{abc_XXX} functions}

\item{truth}{a named numeric vector of known parameter values}

\item{...}{passed on to methods}
}
\value{
a plot of the density functions by wave
}
\description{
Plot the evolution of the density function by wave for SMC and adaptive ABC
}
\examples{

plot_evolution(
  example_adaptive_fit(),
  example_truth()
)

}
\concept{workflow}
