% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-convergence.R
\name{plot_simulations}
\alias{plot_simulations}
\title{Spaghetti plot of resampled posterior fits}
\usage{
plot_simulations(obsdata, fit, sim_fn, method = "auto", n = 200, ...)
}
\arguments{
\item{obsdata}{The observational data. The data in this will typically
be a named list, but could be anything, e.g. dataframe. It is the reference
data that the simulation model is aiming to replicate.}

\item{fit}{A S3 \code{abc_fit} object as output by the \code{abc_XXX} functions}

\item{sim_fn}{a user defined function that takes a set of parameters named
the same as \code{priors_list}. It must return a simulated data set in the
same format as \code{obsdata}, or that can be compared to \code{simdata} by
\code{scorer_fn}. This function must not refer to global parameters, and will be
automatically crated with \code{carrier}.}

\item{method}{one of \code{"auto"}, \code{"count"}, or \code{"density"}, or a named vector
one for each component of \code{obsdata}.}

\item{n}{the number of simulations to plot}

\item{...}{
  Named arguments passed on to \code{\link[=posterior_resample]{posterior_resample}}\describe{
    \item{\code{n_resamples}}{the number of resamples for each parameter combination.}
\item{\code{max_samples}}{the maximum total number of resamples to pick.}
}}
}
\value{
a patchwork of \code{ggplot}s
}
\description{
Spaghetti plot of resampled posterior fits
}
\examples{

plot_simulations(
  example_obsdata(),
  example_adaptive_fit(),
  example_sim_fn
)

}
\concept{workflow}
