% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{pnull}
\alias{pnull}
\title{Null distributions always returns NA}
\usage{
pnull(q, ...)
}
\arguments{
\item{q}{vector of quantiles}

\item{...}{not used}
}
\description{
Null distributions always returns NA
}
\examples{
pnull(c(0.25,0.5,0.75), 0.5, 0.25)
}
\concept{distributions}
