% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-workflow.R
\name{posterior_distance_metrics}
\alias{posterior_distance_metrics}
\title{Generate a set of metrics from component scores}
\usage{
posterior_distance_metrics(posteriors_df, obsscores = NULL, keep_data = FALSE)
}
\arguments{
\item{posteriors_df}{a dataframe of posteriors that have been selected by ABC
this may include columns for scores, weight and/or simulation outputs (
\code{abc_component_score}, \code{abc_summary_distance}, \code{abc_weight}, \code{abc_simulation}
) as well as columns matching the \code{priors} input specification.}

\item{obsscores}{Summary scores for the observational data. This will
be a named list, and is equivalent to the output of \code{scorer_fn},
on the observed data. If not given typically it will be assumed to be all
zeros.}

\item{keep_data}{mainly for internal use this flag gives you the component
scores as a matrix}
}
\value{
a list containing the following items:
\itemize{
\item \code{obsscores}: the input reference scores for each component
\item \code{means}, \code{sds}: the means and sds of each score component
\item \code{cov}: the covariance matrix for the scores
\item \code{mad}: the mean absolute differences between the \code{simscores} and the \code{obsscores}
\item \code{rmsd}: the root mean squared differences between the \code{simscores} and the \code{obsscores}
\item \code{scoreweights}: a the \code{sds} divided by the \code{rmsd}. This weight should
mean that the weights of the  individual summary scores have similar influence
in the overall \code{abc_summary_distance} output once combined during SMC and
adaptive waves, especially if euclidean distances are involved.
\item \code{simscores}: (if \code{keep_data}) a matrix of all the scores from these input
simulation posteriors
\item \code{deltascores}: (if \code{keep_data})  a matrix of the differences between \code{simscores} and \code{obsscores}.
}
}
\description{
The component scores are summary statistics output by the user supplied
\code{scorer_fn} as a named list. These can be variable in scale and location and
various options exist for combining them. They may need to be weighted by
scale as well as importance to get a model that works well. Such weights can
be input into the ABC algortihms using the \code{scoreweights} parameter. This
function helps provide diagnostics for calibrating the \code{scoreweights}
parameter.
}
\examples{

fit = example_rejection_fit()
metrics = posterior_distance_metrics(fit$posteriors)
# other elements available:
metrics$scoreweights
}
\concept{workflow}
