% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-workflow.R
\name{posterior_fit_analytical}
\alias{posterior_fit_analytical}
\title{Fit analytical distribution to posterior samples for generating more waves}
\usage{
posterior_fit_analytical(posteriors_df, priors_list)
}
\arguments{
\item{posteriors_df}{a dataframe of posteriors that have been selected by ABC
this may include columns for scores, weight and/or simulation outputs (
\code{abc_component_score}, \code{abc_summary_distance}, \code{abc_weight}, \code{abc_simulation}
) as well as columns matching the \code{priors} input specification.}

\item{priors_list}{a named list of priors specified as a \code{abc_prior} S3 object
(see \code{priors()}), this can include derived values as unnamed 2-sided
formulae, where the LHS of the formula will be assigned to the value of the
RHS, plus optionally a set of constraints as one sided formulae where the
RHS of the formulae will resolve to a boolean value.}
}
\value{
an \code{abc_prior} S3 object approximating the distribution of the
posterior samples from the same family as the prior.
}
\description{
This function allows "updating" of the prior with a posterior
distribution from the same family as the prior with updated parameters.
}
\details{
This takes weighted posterior samples \eqn{\{(\theta^{(i)}, w^{(i)})\}}
for each parameter \eqn{\theta_j} and fits an analytical distribution
function \eqn{Q_j(\theta_j)} that approximates the posterior marginal for
that parameter.

The resulting empirical distribution \eqn{Q_j} is a \code{dist_fns} object that
includes the support constraints from the original prior. The set of all fitted
marginal distributions \eqn{\{Q_j\}} forms the new proposal list for the next
wave.

Additionally, the weighted covariance matrix \eqn{R} of the samples in the
MVN space (defined by the prior copula) is calculated:
\deqn{
  R = \text{Cov}_{w}(\Phi^{-1}(P_1(\theta_1)), \dots, \Phi^{-1}(P_d(\theta_d)))
}
where \eqn{\Phi^{-1}} is the quantile function of the standard normal.
This covariance matrix is stored as an attribute (\code{"cor"}) of the returned
proposal list and is used to induce correlation structure when sampling
new proposals from the empirical distributions in subsequent waves.
}
\examples{

fit = example_smc_fit()
proposals = posterior_fit_analytical(fit$posteriors, fit$priors)

proposals

}
\concept{workflow}
