% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-workflow.R
\name{posterior_resample}
\alias{posterior_resample}
\title{Generate a set of samples from selected posteriors}
\usage{
posterior_resample(
  posteriors_df,
  sim_fn,
  n_resamples = 1,
  seed = NULL,
  parallel = FALSE,
  max_samples = 200
)
}
\arguments{
\item{posteriors_df}{a dataframe of posteriors that have been selected by ABC
this may include columns for scores, weight and/or simulation outputs (
\code{abc_component_score}, \code{abc_summary_distance}, \code{abc_weight}, \code{abc_simulation}
) as well as columns matching the \code{priors} input specification.}

\item{sim_fn}{a user defined function that takes a set of parameters named
the same as the list \code{priors}. It must return a simulated data set in the
same format as \code{obsdata}, or that can be compared to \code{simdata} by
\code{scorer_fn}. This function must not refer to global parameters, and will be
automatically crated with \code{carrier}.}

\item{n_resamples}{the number of resamples for each parameter combination.}

\item{seed}{an optional random seed}

\item{parallel}{parallelise the simulation? If this is set to true then the
simulation step will be parallelised using \code{furrr}. For this to make any
difference it must have been set up with the following:
\code{future::plan(future::multisession, workers = parallel::detectCores()-2)}}

\item{max_samples}{the maximum total number of resamples to pick.}
}
\value{
a dataframe of the posteriors with an \code{abc_sim} list column
containing the output of \code{sim_fn} called with the parameters in that row.
}
\description{
Once an ABC model fitting is complete the simulation data is generally
only one possible realisation of the parameters, which has been selected for
closeness. To properly compare the output with the observed data we need a
set of posterior re-samples, which are selected from posteriors according to
importance.
}
\examples{

fit = example_adaptive_fit()

sample_df = posterior_resample(
  fit$posteriors,
  sim_fn = example_sim_fn
)

# the fitted simulations are in the `abc_sim` column
sim1 = sample_df$abc_sim[[1]]

sim1 \%>\% lapply(head, 10)

}
\concept{workflow}
