% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-workflow.R
\name{posterior_summarise}
\alias{posterior_summarise}
\title{Calculate a basket of summaries from a weighted list of posterior samples}
\usage{
posterior_summarise(
  posteriors_df,
  priors_list,
  p = c(0.025, 0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95, 0.975)
)
}
\arguments{
\item{posteriors_df}{a dataframe of posteriors that have been selected by ABC
this may include columns for scores, weight and/or simulation outputs (
\code{abc_component_score}, \code{abc_summary_distance}, \code{abc_weight}, \code{abc_simulation}
) as well as columns matching the \code{priors} input specification.}

\item{priors_list}{a named list of priors specified as a \code{abc_prior} S3 object
(see \code{\link{priors()}}), this can include derived values as unnamed 2-sided
formulae, where the LHS of the formula will be assigned to the value of the
RHS, plus optionally a set of constraints as one sided formulae where the
RHS of the formulae will resolve to a boolean value.}

\item{p}{a \code{progressr} progress bar}
}
\value{
a dataframe indexed by parameter with useful summary metrics.
}
\description{
Calculate a basket of summaries from a weighted list of posterior samples
}
\examples{

fit = example_adaptive_fit()
summ = posterior_summarise(fit$posteriors, fit$priors)

summ \%>\% dplyr::glimpse()

}
\concept{workflow}
