% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{print.link_fns}
\alias{print.link_fns}
\title{Print a \code{link_fns} S3 object}
\usage{
\method{print}{link_fns}(x, ...)
}
\arguments{
\item{x}{a \code{link_fns} S3 object}

\item{...}{passed onto methods}
}
\value{
nothing
}
\description{
Boilerplate S3 class operation
}
\concept{link_fns_s3}
\keyword{internal}
