% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-abc_prior-class.R
\name{priors}
\alias{priors}
\title{Construct a set of priors}
\usage{
priors(...)
}
\arguments{
\item{...}{a list of formulae. Two sided will be interpreted as distribution
or derived value specifications. One sided as constraints between parameters.
A distribution is specified as the name of the family of statistical distributions
and their parameters: e.g.: \code{x ~ norm(mean=3,sd=2)}. The name will be matched
to the first hit on the search path.}
}
\value{
an S3 object of class \code{abc_prior} which contains
\itemize{
\item a list of \code{dist_fns}
\item a \code{cor} attribute describing their correlation
\item a \code{derived} attribute describing derive values
\item a \code{constraints} attribute listing the constraints
\item a \code{params} attribute listing the names of the parameters
}
}
\description{
\code{abc_prior} S3 objects are used to hold the specification of prior and
intermediate proposal distributions. They are inputs to the main \code{abc_...()}
workflow functions.
}
\examples{
p = priors(
  mean ~ tidyabc::rgamma2(4,2),
  sd ~ gamma2(2,1),
  shape ~ mean^2 / sd^2,
  rate ~ mean / sd^2,
  ~ mean > sd
)

print(p)

# Plot methods are also provided:
if (interactive()) plot(p)

# constraints:
p@constraints

}
\concept{workflow}
