% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{qlogitnorm}
\alias{qlogitnorm}
\title{Logit-normal distribution}
\usage{
qlogitnorm(p, meanlogit = 0, sdlogit = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of probabilities}

\item{meanlogit}{the mean on the logit scale}

\item{sdlogit}{the sd on the logit scale}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X<=x]}
otherwise \code{P[X>x]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
a vector of probabilities, quantiles, densities or samples.
}
\description{
The logit-normal distribution has a support of 0 to 1.
}
\examples{
qlogitnorm(c(0.25,0.5,0.75), 0, 1)
}
\concept{distributions}
