% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{qlogitnorm2}
\alias{qlogitnorm2}
\title{Logit-normal distribution}
\usage{
qlogitnorm2(
  p,
  prob.0.5 = 0.5,
  kappa = 1 - exp(-1),
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{p}{vector of probabilities}

\item{prob.0.5}{the median on the true scale}

\item{kappa}{a dispersion parameter from 0 (none) to 1 maximum dispersion}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X<=x]}
otherwise \code{P[X>x]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
a vector of probabilities, quantiles, densities or samples.
}
\description{
The logit-normal distribution has a support of 0 to 1.
}
\examples{
qlnorm2(c(0.25,0.5,0.72), 2, 1)
}
\concept{distributions}
