% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{qnull}
\alias{qnull}
\title{Null distributions always returns NA}
\usage{
qnull(p, ...)
}
\arguments{
\item{p}{vector of probabilities}

\item{...}{not used}
}
\description{
Null distributions always returns NA
}
\examples{
qnull(c(0.25,0.5,0.75), 0.5, 0.25)
}
\concept{distributions}
