% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{rcgamma}
\alias{rcgamma}
\title{Sampling: gamma distribution constrained to have mean > sd}
\usage{
rcgamma(n, mean, kappa = 1/mean)
}
\arguments{
\item{n}{number of observations}

\item{mean}{the mean value on the true scale (vectorised)}

\item{kappa}{a coefficient of variation. where 0 is no variability and 1 is
maximally variability (vectorised)}
}
\value{
\code{dgamma} gives the density,
  \code{pgamma} gives the distribution function,
  \code{qgamma} gives the quantile function, and
  \code{rgamma} generates random deviates.

  Invalid arguments will result in return value \code{NaN}, with a warning.

  The length of the result is determined by \code{n} for
  \code{rgamma}, and is the maximum of the lengths of the
  numerical arguments for the other functions.

  The numerical arguments other than \code{n} are recycled to the
  length of the result.  Only the first elements of the logical
  arguments are used.
}
\description{
The following conversion describes the parameters mean and kappa
}
\details{
\deqn{
\text{shape:} \alpha = \frac{1}{\kappa} \\
\text{rate:} \beta = \frac{1}{\mu \times \kappa} \\
\text{scale:} \sigma = \mu \times \kappa \\
}
}
\examples{
rcgamma(10, 2, 0.5)
}
\seealso{
\code{\link[stats:GammaDist]{stats::rgamma()}}
}
\concept{distributions}
