% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{rexpgrowth}
\alias{rexpgrowth}
\title{Randomly sample incident times in an exponentially growing process}
\usage{
rexpgrowth(n, r, t_end, t_start = 0)
}
\arguments{
\item{n}{the number of items to sample}

\item{r}{and exponential growth rate (per unit time)}

\item{t_end}{the end of the observation period}

\item{t_start}{the start of the observation period}
}
\value{
a vector of \code{n} samples from an exponential growth process
}
\description{
Randomly sample incident times in an exponentially growing process
}
\examples{
graphics::hist(rexpgrowth(1000,0.1,40), breaks=40)
graphics::hist(rexpgrowth(1000,-0.1,40), breaks=40)
}
\concept{distributions}
