% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{rexpgrowthI0}
\alias{rexpgrowthI0}
\title{Randomly sample incident times in an exponentially growing process with initial case load}
\usage{
rexpgrowthI0(I0, r, t_end, t_start = 0)
}
\arguments{
\item{I0}{the expected number of cases observed in the first day}

\item{r}{and exponential growth rate (per unit time)}

\item{t_end}{the end of the observation period}

\item{t_start}{the start of the observation period}
}
\value{
a vector of \code{n} samples from an exponential growth process
}
\description{
Randomly sample incident times in an exponentially growing process with initial case load
}
\examples{
hist(rexpgrowthI0(10,0.1,20), breaks=40)
hist(rexpgrowthI0(1000,-0.1,40), breaks=40)
}
\concept{distributions}
