% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{rlogitnorm}
\alias{rlogitnorm}
\title{Logit-normal distribution}
\usage{
rlogitnorm(n, meanlogit = 0, sdlogit = 1)
}
\arguments{
\item{n}{number of observations}

\item{meanlogit}{the mean on the logit scale}

\item{sdlogit}{the sd on the logit scale}
}
\value{
a vector of probabilities, quantiles, densities or samples.
}
\description{
The logit-normal distribution has a support of 0 to 1.
}
\examples{
rlogitnorm(10, 0, 1)
}
\concept{distributions}
