% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{rlogitnorm2}
\alias{rlogitnorm2}
\title{Logit-normal distribution}
\usage{
rlogitnorm2(n, prob.0.5 = 0.5, kappa = 1 - exp(-1))
}
\arguments{
\item{n}{number of observations}

\item{prob.0.5}{the median on the true scale}

\item{kappa}{a dispersion parameter from 0 (none) to 1 maximum dispersion}
}
\value{
a vector of probabilities, quantiles, densities or samples.
}
\description{
The logit-normal distribution has a support of 0 to 1.
}
\examples{
mean(rlogitnorm2(10000,0.75,0.2))
}
\concept{distributions}
