% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{rnull}
\alias{rnull}
\title{Null distributions always returns NA}
\usage{
rnull(n, ...)
}
\arguments{
\item{n}{number of observations}

\item{...}{not used}
}
\description{
Null distributions always returns NA
}
\examples{
rnull(3, c(0.1,0.5,0.9),0.1)
}
\concept{distributions}
