% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{s3_dist_fns}
\alias{s3_dist_fns}
\title{Create a \code{dist_fns} S3 object}
\arguments{
\item{x}{a \code{dist_fns} S3 object}

\item{...}{passed onto methods}
}
\description{
A class wrapping a single (or set) of parametrised distributions and
provides access to the quantile, cumulative probability and random functions
of that specific distribution. Parametrisation is handled on construction.
}
\details{
\code{dist_fns} and \code{dist_fns_list} objects support \code{$} access for fields and \code{@} access
for attributes. \code{dist_fns_list}s can be made with the \code{c()} or \code{rep()} functions,
or with the \code{purrr} style map functions, and they support subsetting.
Individual \code{dist_fns} members of \code{dist_fns_list}s can be accessed with \code{[[}.
}
\keyword{internal}
