% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{s3_dist_fns_common}
\alias{s3_dist_fns_common}
\title{\code{dist_fns} S3 Class operations}
\arguments{
\item{x}{a \code{dist_fns} S3 object}

\item{...}{passed onto methods}
}
\description{
Boilerplate S3 class operation
}
\keyword{internal}
