% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{s3_link_fns}
\alias{s3_link_fns}
\title{Create a \code{link_fns} S3 object}
\arguments{
\item{x}{a \code{link_fns} S3 object}

\item{...}{passed onto methods}
}
\description{
The link function class allows forwards and backwards transformation.
Link functions can be defined by name or using a statistical distribution
in which case the forward link is a logit of the cumulative probability
and the reverse is the quantile of the expit.
}
\details{
\code{link_fns} and \code{link_fns_list} objects support \code{$} access for fields and \code{@} access
for attributes. \code{link_fns_list}s can be made with the \code{c()} or \code{rep()} functions,
or with the \code{purrr} style map functions, and they support subsetting.
Individual \code{link_fns} members of \code{link_fns_list}s can be accessed with \code{[[}.
}
\keyword{internal}
