% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{s3_link_fns_common}
\alias{s3_link_fns_common}
\title{\code{link_fns} S3 Class operations}
\arguments{
\item{x}{a \code{link_fns} S3 object}

\item{...}{passed onto methods}
}
\description{
Boilerplate S3 class operation
}
\keyword{internal}
