% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{s3_link_fns_list}
\alias{s3_link_fns_list}
\title{Manipulate \code{link_fns} S3 object lists}
\arguments{
\item{x}{a \code{link_fns_list} S3 object}

\item{...}{passed onto methods}
}
\description{
These boilerplate functions allow generic list behaviour from \code{link_fns_list}
classes allowing \code{link_fns} S3 objects to be used in lists or dataframes.
}
\keyword{internal}
