% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-sigmoid.R
\name{scale_probability}
\alias{scale_probability}
\title{Scale probabilities by odds ratios}
\usage{
scale_probability(p, odds_ratio)
}
\arguments{
\item{p}{A vector of probabilities}

\item{odds_ratio}{a vector of odds ratios}
}
\value{
an adjusted vector of probabilities
}
\description{
\deqn{
  O = \frac{p}{1-p} \\
  p = \frac{O}{1+O} \\
  p_k = \frac{\frac{kp}{1-p}}{( 1 + \frac{kp}{1-p})} \\
  p_k = \frac{kp}{1 - p + kp}
}
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

scale_probability(0.5, c(0.25,0.5,1,2,4))
}\if{html}{\out{</div>}}
}

