% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-sigmoid.R
\name{sigmoid_params}
\alias{sigmoid_params}
\title{Sigmoid functions}
\arguments{
\item{n}{a vector of lengths}

\item{n_50}{the size at 50 percent}

\item{n_01}{the size at 1 percent, and the limit of the returned \code{p} value
such that \eqn{p \times n \ge n_{01}}}

\item{n_100}{an offset under which the value is 1 (default is 0)}

\item{p_max}{the maximum value at \code{n <= n_100}}
}
\description{
Sigmoid functions
}
\keyword{internal}
