% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_outbreak}
\alias{sim_outbreak}
\title{The \code{sim_outbreak} dataset}
\format{
An object of class \code{list} of length 3.
}
\source{
\url{https://ai4ci.github.io/ggoutbreak/articles/simulation-test-models.html#line-list-simulations}
}
\usage{
sim_outbreak
}
\description{
This is a generated data set of an outbreak based on a branching process
model with fixed parameters. There is a both the full infection line list
as it happened in the simulation and the subset of this data that might have
been observed in a contact tracing exercise where identification is done
through symptoms.
}
\details{
\subsection{\code{sim_outbreak} is a named list with 3 items}{

\tabular{lll}{
\bold{Item} \tab \bold{Type} \tab \bold{Description} \cr
\code{parameters} \tab \code{list[dbl]} \tab the ground truth of the simulation parameters \cr
\code{outbreak_truth} \tab \code{df[outbreak_truth]}* \tab the full infection line list \cr
\code{contact_tracing} \tab \code{df[contact_tracing]}* \tab the "observed" contact tracing subset \cr
}
}

\subsection{\code{df[outbreak_truth]} dataframe with 663 rows and 11 columns}{

The simulation details

\tabular{lll}{
\bold{Column} \tab \bold{Type} \tab \bold{Description} \cr
\code{time} \tab \code{dbl} \tab The true time of infection \cr
\code{id} \tab \code{int} \tab Person unique id \cr
\code{generation_interval} \tab \code{dbl} \tab The time since infector's infection \cr
\code{infector} \tab \code{int} \tab The unique id of the infector \cr
\code{generation} \tab \code{dbl} \tab Which generation is this infection since the simulation start \cr
\code{symptom} \tab \code{lgl} \tab Did this person experience symptoms \cr
\code{symptom_delay} \tab \code{dbl} \tab How long after infection were their symptoms?\cr
\code{symptom_time} \tab \code{dbl} \tab When? (from the simulation start)\cr
\code{observation} \tab \code{lgl} \tab Was this person detected (only if symptoms) \cr
\code{observation_delay} \tab \code{dbl} \tab How long after symptoms were they observed? \cr
\code{observation_time} \tab \code{dbl} \tab When? (from the simulation start) \cr
}
}

\subsection{\code{df[contact_tracing]} dataframe with 663 rows and 4 columns}{

A minimal set of data that might be collected in a contact tracing
exercise.

\tabular{lll}{
\bold{Column} \tab \bold{Type} \tab \bold{Description} \cr
\code{id} \tab \code{int} \tab Unique person id \cr
\code{contact_id} \tab \code{int} \tab Unique id of infectious contact \cr
\code{onset_time} \tab \code{int} \tab Time of symptom onset (from the simulation start) \cr
\code{obs_time} \tab \code{int} \tab Time of observation (from the simulation start) \cr
}
}
}
\keyword{datasets}
