% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{sort.dist_fns_list}
\alias{sort.dist_fns_list}
\title{Sort a \code{dist_fns_list}}
\usage{
\method{sort}{dist_fns_list}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{a \code{dist_fns_list} S3 object}

\item{decreasing}{reverse the sort order}

\item{...}{passed onto methods}
}
\value{
an ordered \code{dist_fns_list} S3 object
}
\description{
These boilerplate functions allow generic list behaviour from \code{dist_fns_list}
classes allowing \code{dist_fns} S3 objects to be used in lists or dataframes.
}
\concept{dist_fns_s3}
\keyword{internal}
