% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-dist_fns-class.R
\name{[<-.dist_fns_list}
\alias{[<-.dist_fns_list}
\title{Assign a subset to a \code{dist_fns_list}}
\usage{
\method{[}{dist_fns_list}(x, ...) <- value
}
\arguments{
\item{x}{a \code{dist_fns_list} S3 object}

\item{...}{passed onto methods}

\item{value}{the value as \code{dist_fns_list} or \code{dist_fns} S3 objects}
}
\value{
the updated \code{dist_fns_list} S3 object
}
\description{
These boilerplate functions allow generic list behaviour from \code{dist_fns_list}
classes allowing \code{dist_fns} S3 objects to be used in lists or dataframes.
}
\concept{dist_fns_s3}
\keyword{internal}
