% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-euclidean.R
\name{summary_distance}
\alias{summary_distance}
\title{Combine simulation scores by euclidean distance.}
\usage{
summary_distance(
  component_scores,
  obsscores = NULL,
  distance_method = c("euclidean", "manhattan", "mahalanobis"),
  wave1_cov = NULL
)
}
\arguments{
\item{component_scores}{\itemize{
\item a list column of scores. each entry is itself a list
}}

\item{obsscores}{Summary scores for the observational data. This will
be a named list, and is equivalent to the output of \code{scorer_fn},
on the observed data. If not given typically it will be assumed to be all
zeros.}

\item{distance_method}{what metric is used to combine \code{simscores} and \code{obsscores}
and is one of \code{"euclidean"}, \code{"manhattan"}, or \code{"mahalanobis"}.}
}
\value{
a vector of distances (with a \code{cov} attribute of the wave 1
covariance matrix)
}
\description{
calculate the overall score for an individual
simulation, based on the components output by \code{score_list}. This function
is applied to a list of simulation scores and combines them into a
single metric by comparing them to the scores of the original observed data
}
\details{
TODO: needs refactoring to work with whole sets of scores rather than
individual ones, so we can do a mahalabois distance.
}
\keyword{internal}
