% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{test_simulation}
\alias{test_simulation}
\title{Run the simulation for one set of parameters}
\usage{
test_simulation(sim_fn, scorer_fn, ..., seed = NULL)
}
\arguments{
\item{sim_fn}{a user defined function that takes a set of parameters named
the same as the list \code{priors}. It must return a simulated data set in the
same format as \code{obsdata}, or that can be compared to \code{simdata} by
\code{scorer_fn}. This function must not refer to global parameters, and will be
automatically crated with \code{carrier}.}

\item{scorer_fn}{a user supplied function that matches the following
signature \code{scorer_fn(simdata, obsdata, ....)}, i.e. it takes data in the
format of \code{simdata} paired with the original \code{obsdata} and returns a named
list of scores per simulation. This function can make use of the
\verb{calculate_*()} set of functions to compare components of the simulation to
the original data. This function must not refer to global parameters, and
will be automatically crated with \code{carrier}. If this is a purrr style
function then \code{.x} will refer to simulation output and \code{.y} to original
observation data.}

\item{...}{simulation parameters, must be named}

\item{seed}{an optional random seed}
}
\value{
a list containing the parameters as \code{truth} and an
instance of the simulation as \code{obsdata}
}
\description{
Run the simulation for one set of parameters
}
\concept{workflow}
