% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-transform.R
\name{transform}
\alias{transform}
\title{Generate a distribution from a link transform of another}
\usage{
transform(link, dist, ..., name = NULL)
}
\arguments{
\item{link}{a link function (or name of a link function)}

\item{dist}{distribution(s) as a name, function or a \code{dist_fn} S3 object}

\item{...}{parameters for the underlying distribution if \code{dist} is a name or
function.}

\item{name}{a name for the link function}
}
\value{
a \code{dist_fn} or \code{dist_fn_list} holding the transformed distribution(s)
}
\description{
Generates a new distribution by applying a link transformation to an existing
distribution \code{dist}. If \eqn{X \sim \text{dist}} and \eqn{h} is the link function,
this function returns the distribution of \eqn{Y = h^{-1}(X)}.
The CDF \eqn{F_Y}, quantile function \eqn{Q_Y}, PDF \eqn{f_Y}, and RNG \eqn{R_Y}
of the transformed distribution are derived from the original distribution's
functions \eqn{F_X}, \eqn{Q_X}, \eqn{f_X}, \eqn{R_X} and the link function \eqn{h}
and its inverse \eqn{h^{-1}} as follows:
\deqn{
F_Y(y) = F_X(h(y))
}
\deqn{
Q_Y(p) = h^{-1}(Q_X(p))
}
\deqn{
f_Y(y) = f_X(h(y)) \cdot |h'(y)|
}
\deqn{
R_Y(n) = h^{-1}(R_X(n))
}
where \eqn{h'(y)} is the derivative of the link function. This function implements
these transformations for the \code{p}, \code{q}, \code{d}, and \code{r} functions of the resulting
\code{dist_fns} object.
}
\examples{

n = as.dist_fns("norm",mean=0.5, sd=0.1)
t = transform("log",n)

plot(t)+ggplot2::geom_function(fun=~ dlnorm(.x, 0.5, 0.1), linetype="dashed")

}
\concept{empirical}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{



t = transform("log","norm")
ps = seq(0,1,0.1)
qs = 0:6
testthat::expect_equal(t$q(ps),qlnorm(ps))
testthat::expect_equal(t$p(qs),plnorm(qs))
testthat::expect_equal(t$d(qs),dlnorm(qs))

t2 = transform("log","norm", 0.4, 0.1)
testthat::expect_equal(t2$q(ps),qlnorm(ps,0.4,0.1))

}\if{html}{\out{</div>}}
}

