% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{type_sum.link_fns}
\alias{type_sum.link_fns}
\title{Pillar type name}
\usage{
\method{type_sum}{link_fns}(x, ...)
}
\arguments{
\item{x}{a \code{link_fns} S3 object}

\item{...}{passed onto methods}
}
\value{
a string
}
\description{
Boilerplate S3 class operation
}
\concept{link_fns_s3}
\keyword{internal}
