% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-wasserstein.R
\name{wasserstein_calculator}
\alias{wasserstein_calculator}
\title{Generate a function to calculate a wasserstein distance}
\usage{
wasserstein_calculator(obs, debias = FALSE, bootstraps = 1)
}
\arguments{
\item{obs}{A vector of observations}

\item{debias}{Should the simulations be shifted to match the mean of the
observed data}

\item{bootstraps}{Randomly resample from the simulated data points to match
the observed size this many times and combine the output by averaging. The
alternative, when this is 1 (the default) matches the sizes by selecting
and/or repeating the simulated data points in order (deterministically)}
}
\value{
a function that takes parameter \code{sim} and returns a
length normalised wasserstein distance. This is the average distance an
individual data point must be shifted to match the reference data normalised
by the average distance of the reference data from the mean. The function
also takes a \code{p} parameter which can be a \code{progressr} progress bar which must
be named.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This function takes reference data in the forms of individual observations in
the form of for example event times and returns a crated function to
calculate the wasserstein distance from simulated data to the observed data.
For large simulations this will be quicker than \code{calculate_wasserstein} but
must be used with a crated \code{scorer_fn}

In the comparison unequal lengths of the data can be accommodated. The
simulated data is recycled, and sampled, until the same length as the
observed data before the comparison.
}
\concept{workflow}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

tmp = wasserstein_calculator(0:10)

# zero if no distance
testthat::expect_equal(tmp(0:10), 0)
testthat::expect_equal(tmp(10:0), 0)


}\if{html}{\out{</div>}}
}

