% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-empirical.R
\name{wbw.nrd}
\alias{wbw.nrd}
\title{Weighted bandwidth selector}
\usage{
wbw.nrd(x, w)
}
\arguments{
\item{x}{data}

\item{w}{weights}
}
\value{
a bandwidth based on weighted data
}
\description{
Weighted bandwidth selector
}
\concept{empirical}
