% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-empirical.R
\name{wmean}
\alias{wmean}
\title{Weighted mean}
\usage{
wmean(x, w = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{either a vector of samples from a distribution \code{X} or cut-offs for
cumulative probabilities when combined with \code{p}}

\item{w}{for data fits, a vector the same length as \code{x} giving the
importance weight of each observation. This does not need to be normalised.
There must be some non zero weights, and all must be finite.}

\item{na.rm}{remove NAs (default TRUE)}
}
\value{
a standard deviation
}
\description{
a simple alias for base \verb{weighted,mean}
}
\examples{
#' # unweighted:
wmean(x = stats::rnorm(1000))

# weighted:
wmean(x = seq(-2,2,0.1), w = stats::dnorm(seq(-2,2,0.1)))

}
\concept{empirical}
