% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-empirical.R
\name{wquantile}
\alias{wquantile}
\title{Quantile from weighted data with link function support}
\usage{
wquantile(p, x, w = NULL, link = "identity", names = TRUE, window = 7)
}
\arguments{
\item{p}{the probabilities for which to estimate quantiles from the data}

\item{x}{a set of observations}

\item{w}{for data fits, a vector the same length as \code{x} giving the
importance weight of each observation. This does not need to be normalised.
There must be some non zero weights, and all must be finite.}

\item{link}{a link function. Either as name, or a \code{link_fns} S3 object. In
the latter case this could be derived from a statistical distribution by
\verb{as.link_fns(<dist_fns>)}. This supports the use of a prior to define the
support of the empirical function, and is designed to prevent tail
truncation. Support for the updated quantile function will be the same as
the provided prior.}

\item{names}{name the resulting quantile vector}

\item{window}{the number of data points to include when estimating the
quantile. The closest \code{window} points are picked and used as a distance
weighted linear interpolation of the weighted CDF in logit-link space. This
tends to give good results for extrapolating tails.}
}
\value{
a vector of quantiles
}
\description{
This quantile function has different order of parameters from
base quantile. It takes a weight and a link function specification which
allows us to define the support of the quantile function. It is
optimised for imputing the tail of distributions and not speed.
}
\details{
This is a moderately expensive function to call (in memory terms), as it
needs to construct the whole quantile function. if there are multiple calls
consider using \code{empirical()} to build a quantile function and using that.
}
\examples{

# unweighted:
wquantile(p = c(0.25,0.5,0.75), x = stats::rnorm(1000))

# weighted:
wquantile(p = c(0.25,0.5,0.75), x = seq(-2,2,0.1), w = stats::dnorm(seq(-2,2,0.1)))

}
\concept{empirical}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{


test = function(rfn,qfn, link,..., n = 100000, tol=1000/(n+10000)) {
  testthat::expect_equal(abs(
   unname(wquantile(c(0.025, 0.5, 0.975),rfn(100000,...),link=link,names=FALSE)-
    qfn(c(0.025, 0.5, 0.975), ...))
  ), c(0,0,0), tolerance=tol)
}

withr::with_seed(123, {

  test(stats::rnorm,stats::qnorm,"identity",n = 10000)
  test(stats::rnorm,stats::qnorm,"identity",mean=4,n = 10000)
  test(stats::rnorm,stats::qnorm,"identity",sd=3, n = 100000, tol=0.05)

  test(stats::rnorm,stats::qnorm,"identity",n = 5000)
  test(stats::rnorm,stats::qnorm,"identity",n = 1000)
  test(stats::rnorm,stats::qnorm,"identity",n = 100)
  test(stats::rnorm,stats::qnorm,"identity",n = 30)

  test(stats::rgamma,stats::qgamma,"log", 4,n = 10000)
  test(stats::rgamma,stats::qgamma,"log", 4,n = 5000)
  test(stats::rgamma,stats::qgamma,"log", 4,n = 1000)
  test(stats::rgamma,stats::qgamma,"log", 4, 3,n = 100)
  test(stats::rgamma,stats::qgamma,"log", 4,n = 30)

  test(stats::runif,stats::qunif,as.link_fns(c(0,10)),0,10)

})


}\if{html}{\out{</div>}}
}

