% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-empirical.R
\name{wt_quant}
\alias{wt_quant}
\title{Quantile from weighted data}
\arguments{
\item{p}{the quantile probabilities}

\item{names}{should the result be named}

\item{na.rm}{remove NAs from observations (and weights)}

\item{x}{the observations}

\item{w}{importance weights. Must be some non zero weights, and all must be
finite. They don't need to sum to one.}

\item{link}{a link function mapping observations to infinite support}
}
\description{
Quantile from weighted data
}
\keyword{internal}
