% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-wasserstein.R
\name{calculate_wasserstein}
\alias{calculate_wasserstein}
\title{Calculate a wasserstein distance}
\usage{
calculate_wasserstein(sim, obs, debias = FALSE)
}
\arguments{
\item{sim}{A vector of simulated data points}

\item{obs}{A vector of observed data points}

\item{debias}{Should the simulations be shifted to match the mean of the
observed data}
}
\value{
a length normalised wasserstein distance. This is the average distance an
individual simulated data point must be shifted to match the observed data normalised
by the average distance of the observed data from the mean.
}
\description{
This function takes simulation and observed data and calculates a normalised
wasserstein distance.
}
\details{
In the comparison unequal lengths of the data can be accommodated. The
simulated data is recycled, and sampled, until the same length as the
observed data before the comparison.
}
\concept{workflow}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{


testthat::expect_equal(calculate_wasserstein(0:10, 10:0), 0)

# zero if no distance
testthat::expect_equal(calculate_wasserstein(0:10, 0:10), 0)
testthat::expect_equal(calculate_wasserstein(10:0, 0:10), 0)

# normalised so that all mass at mean = 1
testthat::expect_equal(calculate_wasserstein(rep(5, 11), 0:10), 1)

# smaller sample recycled and normalises to same value
testthat::expect_equal(calculate_wasserstein(rep(5, 5), 0:10), 1)

# should be ((0+1+0+1+0+0+0+1+0+1+0) / 11) / ((5+4+3+2+1+0+1+2+3+4+5) / 11) = 0.1333...
testthat::expect_equal(
  calculate_wasserstein(
    c(0, 0, 2, 2, 4, 5, 6, 8, 8, 10, 10),
    0:10
 ),
  0.133333333333333
)

withr::with_seed(100, {
   ref = rnorm(1000)
   cmp1 = rnorm(1000)
   cmp2 = rnorm(1000, sd=2)
   cmp3 = rnorm(100)
})

testthat::expect_equal(calculate_wasserstein(cmp1,ref), 0.0576417503974498)
testthat::expect_equal(calculate_wasserstein(cmp2,ref), 1.04950621760385)
testthat::expect_equal(calculate_wasserstein(cmp3,ref), 0.212977231452674)

}\if{html}{\out{</div>}}
}

