% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-abc.R
\name{.calculate_weights_smc}
\alias{.calculate_weights_smc}
\title{Calculate weights for particles in a new wave}
\usage{
.calculate_weights_smc(sim_df, epsilon, prev_sim_df = NULL, kernel)
}
\arguments{
\item{sim_df}{the output of a wave of simulation including a
\code{abc_summary_distance} column}

\item{epsilon}{epsilon is a tolerance threshold that controls how closely
simulated summaries must match the observed ones to be considered
plausible. This is in the unit of \code{abc_summary_distance}. Initially the 0.5
quantile of distances, in subsequent waves this might be decreased.
It is the scale parameter of the kernel function. $K_h(|u|)$}

\item{prev_sim_df}{the output of a previous ABC wave including a
\code{abc_weight} column}

\item{kernel}{one of \code{"epanechnikov"} (default), \code{"uniform"}, \code{"triangular"},
\code{"biweight"}, or \code{"gaussian"}. The kernel defines how the distance metric
translates into the importance weight that decides whether a given
simulation and associated parameters should be rejected or held for the
next round. All kernels except \code{gaussian} have a hard cut-off outside of
which the probability of acceptance of a particle is zero. Use of
\code{gaussian} kernels can result in poor convergence.}
}
\value{
the \code{sim_df} with an \code{abc_weight} column
}
\description{
The ABC weights need to be calculated for sampling from the proposal
distribution. They depend on the priors, an acceptance tolerance and the
weighted particles from the previous wave, which define the proposal
probability distribution.
}
\keyword{internal}
