% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-abc.R
\name{.calculate_weights_wave_one}
\alias{.calculate_weights_wave_one}
\title{Calculate weights for particles in a new wave}
\usage{
.calculate_weights_wave_one(sim_df, epsilon, kernel)
}
\arguments{
\item{sim_df}{the output of a wave of simulation including a
\code{abc_summary_distance} column}

\item{epsilon}{epsilon is a tolerance threshold that controls how closely
simulated summaries must match the observed ones to be considered
plausible. This is in the unit of \code{abc_summary_distance}. Initially the 0.5
quantile of distances, in subsequent waves this might be decreased.
It is the scale parameter of the kernel function. $K_h(|u|)$}

\item{kernel}{one of \code{"epanechnikov"}, \code{"uniform"}, \code{"triangular"}, \code{"biweight"},
or \code{"gaussian"}. The kernel defines how the distance metric translates into
the importance weight that decides whether a given simulation and associated
parameters should be rejected or held for the next round.}
}
\value{
the \code{sim_df} with an \code{abc_weight} column
}
\description{
The ABC weights need to be calculated for sampling from the proposal
distribution. In the situation where we are sampling from the prior.
this is just their distance by a kernel function.
}
\keyword{internal}
