% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{.log_kernel}
\alias{.log_kernel}
\title{Generate a log of a kernel function for ABC weight calculation}
\usage{
.log_kernel(
  u,
  epsilon,
  kernel = c(c("epanechnikov", "uniform", "triangular", "biweight", "gaussian"))
)
}
\arguments{
\item{u}{the distance (dissimilarilty) between}

\item{epsilon}{epsilon is a tolerance threshold that controls how closely
simulated summaries must match the observed ones to be considered
plausible. This is in the unit of \code{abc_summary_distance}. Initially the 0.5
quantile of distances, in subsequent waves this might be decreased.
It is the scale parameter of the kernel function. $K_h(|u|)$}

\item{kernel}{one of "epanechnikov", "uniform", "triangular", "biweight",
"gaussian"}
}
\value{
the log of the kernel function
}
\description{
\deqn{
\text{Uniform: } K(d, \epsilon) = \frac{1}{2\epsilon} \mathbb{I}(d \leq \epsilon)
}
\deqn{
\text{Triangular: } K(d, \epsilon) = \frac{2}{\epsilon}\left(1 - \frac{d}{\epsilon}\right) \mathbb{I}(d \leq \epsilon)
}
\deqn{
\text{Epanechnikov: } K(d, \epsilon) = \frac{3}{4\epsilon}\left(1 - \frac{d^2}{\epsilon^2}\right) \mathbb{I}(d \leq \epsilon)
}
\deqn{
\text{Biweight: } K(d, \epsilon) = \frac{15}{16\epsilon}\left(1 - \frac{d^2}{\epsilon^2}\right)^2 \mathbb{I}(d \leq \epsilon)
}
\deqn{
\text{Gaussian: } K(d, h) = \frac{1}{\sqrt{2\pi h^2}} \exp\left(-\frac{d^2}{2h^2}\right)
}
where \eqn{d} is the distance between observed and simulated summary
statistics, \eqn{\epsilon} is the tolerance level, \eqn{h} is the bandwidth
parameter for the Gaussian kernel, and \eqn{\mathbb{I}} is the indicator
function taking the value 1 if \eqn{d \leq \epsilon} and 0 otherwise (this
applies to all kernels except Gaussian).
}
\keyword{internal}
