% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-abc.R
\name{.sample_priors}
\alias{.sample_priors}
\title{Sample for the prior distribution}
\usage{
.sample_priors(proposal_list, n_sims)
}
\arguments{
\item{proposal_list}{a list of empirical probability distributions that map
MVN space to proposal space, and are the "prior" for each adaptive wave.
This is already used to generate the proposals and their mapping in \code{sim_df}}
}
\value{
a data frame of samples in MVN (prefixed \code{abc_mvn_}) and parameter
space.
}
\description{
This uses a multivariate normal and a copula to generate correlated
structure. The correlation is held as an attribute in the proposals.
N.B. this used to construct derived values and apply constraints but this
was complex and has been deferred. Constraints change the probability
distribution also and will affect validity of proposal distribution.
}
\keyword{internal}
