% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-abc.R
\name{.sample_using_perturbation}
\alias{.sample_using_perturbation}
\title{Generate a new set of particles from a previous wave using perturbation}
\usage{
.sample_using_perturbation(priors_list, n_sims, prev_sim_df)
}
\arguments{
\item{priors_list}{the list of priors as a named list of \code{dist_fn}s, plus one
sided formulae as constraints and 2 sided formulae as derived values.
A correlation matrix may also be present as an attribute \code{cor}.}

\item{n_sims}{The number of simulations to run per wave (for SMC and Adaptive)
or overall (for Rejection). For rejection sampling a large number is
recommended, for the others sma}

\item{prev_sim_df}{the output of a previous ABC wave including a
\code{abc_weight} column}
}
\value{
a \code{sim_df}
}
\description{
This set of particles will be close to existing ones, depending on the
kernel_t parameter.
}
\keyword{internal}
