% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-link_fns-class.R
\name{new_link_fns}
\alias{new_link_fns}
\title{Create a new \code{link_fns} S3 object}
\usage{
new_link_fns(
  trans = ~.x,
  inv = ~.x,
  support = c(-Inf, Inf),
  ddxtrans = NULL,
  ddxinv = NULL,
  name = NULL
)
}
\arguments{
\item{trans}{the forward transforming function}

\item{inv}{the reverse inverting function}

\item{support}{the support or support of the link}

\item{range}{the range of the link (typically \code{-Inf} to \code{Inf})}
}
\value{
a new \code{link_fns} S3 object
}
\description{
Default is an identity link
}
\concept{link_fns_s3}
\keyword{internal}
