% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-convergence.R
\name{plot_correlations}
\alias{plot_correlations}
\title{A parameter posterior correlation plot}
\usage{
plot_correlations(posteriors_df, truth = NULL)
}
\arguments{
\item{posteriors_df}{a dataframe of posteriors that have been selected by ABC
this may include columns for scores, weight and/or simulation outputs (
\code{abc_component_score}, \code{abc_summary_distance}, \code{abc_weight}, \code{abc_simulation}
) as well as columns matching the \code{priors} input specification.}

\item{truth}{a named numeric vector of known parameter values}
}
\value{
a patchwork of ggplots including density and
2d scatters for each combination of posteriors.
}
\description{
A parameter posterior correlation plot
}
\examples{

plot_correlations(
  example_adaptive_fit(),
  example_truth()
)

}
\concept{workflow}
