% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-convergence.R
\name{plot_evolution}
\alias{plot_evolution}
\title{Plot the evolution of the density function by wave for SMC and adaptive ABC}
\usage{
plot_evolution(fit, truth = NULL, ...)
}
\arguments{
\item{fit}{A S3 \code{abc_fit} object as output by the \code{abc_XXX} functions}

\item{truth}{a named numeric vector of known parameter values}

\item{...}{passed on to methods
  Named arguments passed on to \code{\link[=plot.dist_fns_list]{plot.dist_fns_list}}\describe{
    \item{\code{mapping}}{override default aesthetics with \code{name}, \code{id} or \code{group}}
\item{\code{steps}}{resolution of the plot}
\item{\code{tail}}{the minimum tail probability to plot}
\item{\code{plot_quantiles}}{by default the quantiles of the distribution are
plotted over the density sometimes this makes it hard to read}
\item{\code{smooth}}{by default some additional smoothing is used to cover up small
discontinuities in the PDF.}
}}
}
\value{
a plot of the density functions by wave
}
\description{
Plot the evolution of the density function by wave for SMC and adaptive ABC
}
\examples{

plot_evolution(
  example_adaptive_fit(),
  example_truth()
)

}
\concept{workflow}
