% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-workflow.R
\name{posterior_distance_metrics}
\alias{posterior_distance_metrics}
\title{Generate a set of metrics from component scores}
\usage{
posterior_distance_metrics(posteriors_df, obsscores = NULL, keep_data = FALSE)
}
\arguments{
\item{posteriors_df}{a dataframe of posteriors that have been selected by ABC
this may include columns for scores, weight and/or simulation outputs (
\code{abc_component_score}, \code{abc_summary_distance}, \code{abc_weight}, \code{abc_simulation}
) as well as columns matching the \code{priors} input specification.}

\item{obsscores}{Summary scores for the observational data. This will
be a named list, and is equivalent to the output of \code{scorer_fn},
on the observed data. If not given typically it will be assumed to be all
zeros.}

\item{keep_data}{mainly for internal use this flag gives you the component
scores as a matrix}
}
\value{
a list containing the following items:
\itemize{
\item \code{obsscores}: the input reference scores for each component
\item \code{means}, \code{sds}: the means and sds of each score component
\item \code{cov}: the covariance matrix for the scores
\item \code{mad}: the mean absolute differences between the \code{simscores} and the \code{obsscores}
\item \code{rmsd}: the root mean squared differences between the \code{simscores} and the \code{obsscores}
\item \code{scoreweights}: a the \code{sds} divided by the \code{rmsd}. This weight should
mean that the weights of the  individual summary scores have similar influence
in the overall \code{abc_summary_distance} output once combined during SMC and
adaptive waves, especially if euclidean distances are involved.
\item \code{simscores}: (if \code{keep_data}) a matrix of all the scores from these input
simulation posteriors
\item \code{deltascores}: (if \code{keep_data})  a matrix of the differences between \code{simscores} and \code{obsscores}.
}
}
\description{
The component scores are summary statistics output by the user supplied
\code{scorer_fn} as a named list. These can be variable in scale and location and
various options exist for combining them. They may need to be weighted by
scale as well as importance to get a model that works well. Such weights can
be input into the ABC algortihms using the \code{scoreweights} parameter. This
function helps provide diagnostics for calibrating the \code{scoreweights}
parameter.
}
\details{
Given a list of component scores \eqn{S_s^{(i)} = (s_1^{(i)}, \dots, s_m^{(i)})}
from \eqn{n} simulations and observed scores \eqn{S_o = (s_1^{(o)}, \dots, s_m^{(o)})},
this function calculates:
\itemize{
\item The mean \eqn{\mu_j} and standard deviation \eqn{\sigma_j} of each score component \eqn{j}:
\deqn{
    \mu_j = \frac{1}{n}\sum_{i=1}^n s_j^{(i)}, \quad
    \sigma_j = \sqrt{\frac{1}{n-1}\sum_{i=1}^n (s_j^{(i)} - \mu_j)^2}
  }
\item The covariance matrix \eqn{\Sigma} of the score components.
\item The root mean squared difference (RMSD) between simulated and observed scores for each component:
\deqn{
    \text{RMSD}_j = \sqrt{\frac{1}{n}\sum_{i=1}^n (s_j^{(i)} - s_j^{(o)})^2}
  }
\item A recommended vector of \code{scoreweights} \eqn{w_j}, calculated as the ratio of the component's
standard deviation to its RMSD, normalized to sum to 1:
\deqn{
    w_j = \frac{\sigma_j / \text{RMSD}_j}{\sum_{k=1}^m (\sigma_k / \text{RMSD}_k)}
  }
These weights help balance the influence of different summary statistics in
the overall distance metric, especially when using Euclidean distance.
}
}
\examples{

fit = example_rejection_fit()
metrics = posterior_distance_metrics(fit$posteriors)

# other elements available but this is the most important and tells you what
# the relative sizes of the component scores from `scorer_fn` in this sample.
# If this is a sample from the prior then this gives us a way to judge the
# most appropriate relative weighting of each component:
metrics$scoreweights
}
\concept{workflow}
