% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-workflow.R
\name{posterior_fit_empirical}
\alias{posterior_fit_empirical}
\title{Fit empirical distribution to posterior samples for generating more waves}
\usage{
posterior_fit_empirical(posteriors_df, priors_list, knots = NULL, bw = 0.1)
}
\arguments{
\item{posteriors_df}{a dataframe of posteriors that have been selected by ABC
this may include columns for scores, weight and/or simulation outputs (
\code{abc_component_score}, \code{abc_summary_distance}, \code{abc_weight}, \code{abc_simulation}
) as well as columns matching the \code{priors} input specification.}

\item{priors_list}{a named list of priors specified as a \code{abc_prior} S3 object
(see \code{\link{priors()}}), this can include derived values as unnamed 2-sided
formulae, where the LHS of the formula will be assigned to the value of the
RHS, plus optionally a set of constraints as one sided formulae where the
RHS of the formulae will resolve to a boolean value.}

\item{knots}{the number of knots to model the CDF with. Optional, and will be
typically inferred from the data size. Small numbers tend to work better if
we expect the distribution to be unimodal.}

\item{bw}{for Adaptive ABC data distributions are smoothed before modelling
the CDF. Over smoothing can reduce convergence, under-smoothing may result
in noisy posterior estimates. This is in units of the ESS and defaults to
0.1.}
}
\value{
an \code{abc_prior} S3 object approximating the distribution of the
posterior samples, and adhering to the support of the provided priors.
}
\description{
This function allows "updating" of the prior with an empirical posterior
distribution which will retain the bounds of the prior, and is effectively a
spline based transform of the prior distribution, based on the density of
data in prior space. This gives a clean density when data is close to a
prior distribution limit and work better than a standard density for
}
\examples{

fit = example_smc_fit()
proposals = posterior_fit_empirical(fit$posteriors, fit$priors)

proposals

}
\concept{workflow}
