% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-empirical.R
\name{skew}
\alias{skew}
\title{Calculate the skew of a set of data}
\usage{
skew(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of observations}

\item{na.rm}{remove \code{NA}s?}
}
\value{
the skew
}
\description{
Calculate the skew of a set of data
}
\examples{
skew(stats::rnorm(1000))
skew(stats::rbeta(1000, 1, 8)) # positively (left) skewed
skew(stats::rbeta(1000, 8, 1)) # negatively (right) skewed

skew(stats::rlnorm(1000))
}
\concept{empirical}
