% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-workflow.R
\name{test_simulation}
\alias{test_simulation}
\title{Run the simulation for one set of parameters}
\usage{
test_simulation(
  sim_fn,
  scorer_fn,
  ...,
  params = NULL,
  obsdata = NULL,
  seed = NULL,
  debug = FALSE,
  debug_errors = !debug
)
}
\arguments{
\item{sim_fn}{a user defined function that takes a set of parameters named
the same as \code{priors_list}. It must return a simulated data set in the
same format as \code{obsdata}, or that can be compared to \code{simdata} by
\code{scorer_fn}. This function must not refer to global parameters, and will be
automatically crated with \code{carrier}.}

\item{scorer_fn}{a user supplied function that matches the following
signature \code{scorer_fn(simdata, obsdata, ....)}, i.e. it takes data in the
format of \code{simdata} paired with the original \code{obsdata} and returns a named
list of component scores per simulation. This function can make use of the
\verb{calculate_*()} set of functions to compare components of the simulation to
the original data. This function must not refer to global parameters, and
will be automatically crated with \code{carrier}. If this is a purrr style
function then \code{.x} will refer to simulation output and \code{.y} to original
observation data.}

\item{...}{simulation parameters, must be named}

\item{params}{a named list of simulation parameters to test (as an alternative
to including them in \code{...})}

\item{obsdata}{The observational data. The data in this will typically
be a named list, but could be anything, e.g. dataframe. It is the reference
data that the simulation model is aiming to replicate.}

\item{seed}{an optional random seed}

\item{debug}{start the simulation function in debug mode. This will step
through both the \code{sim_fn} and the \code{scorer_fn} line by line to check that
the behaviour is as intended.}

\item{debug_errors}{Errors that crop up in \code{sim_fn} during a simulation due
to anomolous value combinations are hard to debug. If this flag is set,
whenever a \code{sim_fn} or \code{scorer_fn} throws an error an interactive debugging
session is started with the failing parameter combinations. This is not
compatible with running in parallel.}
}
\value{
a list containing the parameters as \code{truth} and an
instance of the simulation as \code{obsdata}, \code{obsscores} is the result of
comparing the \code{obsdata} with itself and is usually going to result in
zeros. Both \code{sim_fn} and \code{scorer_fn} are rewritten to make sure that all
package names are fully qualified. The rewritten versions are refurned in
the result list (as \code{sim_fn} and \code{scorer_fn} respectively)
}
\description{
Run the simulation for one set of parameters
}
\concept{workflow}
