% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyabc-package.R
\docType{package}
\name{tidyabc-package}
\alias{tidyabc}
\alias{tidyabc-package}
\title{tidyabc: Approximate Bayesian Computing with Tidy Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A flexible framework for Approximate Bayesian Computation (ABC) that integrates with the tidyverse. Define simulation models and summary statistics as standard R functions, use 'dist_fns' to represent prior and posterior distributions, and perform inference via rejection sampling, Sequential Monte Carlo (SMC), or Adaptive ABC. The package provides tools for diagnostics, visualization, and convergence assessment, enabling reproducible Bayesian inference for complex models with intractable likelihoods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ai4ci.github.io/tidyabc}
  \item \url{https://github.com/ai4ci/tidyabc}
  \item Report bugs at \url{https://github.com/ai4ci/tidyabc/issues}
}

}
\author{
\strong{Maintainer}: Robert Challen \email{rob.challen@bristol.ac.uk} (\href{https://orcid.org/0000-0002-5504-7768}{ORCID})

}
\keyword{internal}
