% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyabc-package.R
\docType{package}
\name{tidyabc-package}
\alias{tidyabc}
\alias{tidyabc-package}
\title{tidyabc: Approximate Bayesian Computing with Tidy Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Robert Challen \email{rob.challen@bristol.ac.uk} (\href{https://orcid.org/0000-0002-5504-7768}{ORCID})

}
\keyword{internal}
