% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyabc-package.R
\name{tidyabc_common}
\alias{tidyabc_common}
\title{Common workflow options}
\arguments{
\item{obsdata}{The observational data. The data in this will typically
be a named list, but could be anything, e.g. dataframe. It is the reference
data that the simulation model is aiming to replicate.}

\item{priors_list}{a named list of priors specified as \code{dist_fns} (see
\code{as.dist_fns()}), plus optionally, derived values as unnamed 2-sided
formulae, where the LHS of the formula will be assigned to the value of the
RHS, plus optionally a set of constraints as one sided formulae where the
RHS of the formulae will resolve to a boolean value.}

\item{sim_fn}{a user defined function that takes a set of parameters named
the same as the list \code{priors}. It must return a simulated data set in the
same format as \code{obsdata}, or that can be compared to \code{simdata} by
\code{scorer_fn}. This function must not refer to global parameters, and will be
automatically crated with \code{carrier}.}

\item{simdata}{The simulated data. The simulated data must be in exactly
the same format as \code{obsdata}, or that can be compared to \code{simdata} by
\code{scorer_fn}.}

\item{scorer_fn}{a user supplied function that matches the following
signature \code{scorer_fn(simdata, obsdata, ....)}, i.e. it takes data in the
format of \code{simdata} paired with the original \code{obsdata} and returns a named
list of scores per simulation. This function can make use of the
\verb{calculate_*()} set of functions to compare components of the simulation to
the original data. This function must not refer to global parameters, and
will be automatically crated with \code{carrier}. If this is a purrr style
function then \code{.x} will refer to simulation output and \code{.y} to original
observation data.}

\item{obsscores}{Summary scores for the observational data. This will
be a named list, and is equivalent to the output of \code{scorer_fn},
on the observed data. If not given typically it will be assumed to be all
zeros.}

\item{simscores}{Scores for the simulated data, as output by \code{scorer_fn}.}

\item{distance_method}{what metric is used to combine \code{simscores} and \code{obsscores}
and is one of \code{"euclidean"}, \code{"manhattan"}, or \code{"mahalanobis"}.}

\item{keep_simulations}{keep the individual simulation results in the output
of an ABC workflow. This can have large implications for the size of the
result. It may also not be what you want and it is probably worth considering
resampling the posteriors rather than keeping the simulations.}

\item{acceptance_rate}{What proportion of simulations to keep in ABC rejection
or hard ABC parts of the algorithms.}

\item{posteriors_df}{a dataframe of posteriors that have been selected by ABC
this may include columns for scores, weight and/or simulation outputs (
\code{abc_component_score}, \code{abc_summary_distance}, \code{abc_weight}, \code{abc_simulation}
) as well as columns matching the \code{priors} input specification.}

\item{kernel_t}{A kernel bandwidth parameter for proposals. This controls the
amount of noise that particles are perturbed by (and hence the spread of
the PDF of the proposal distribution), and 1 is approximately 34\% of the
proposal distribution at the centre. Smaller values (default is 0.2) give a
smaller step size in generating new proposals, and proposals will be closer
to currently accepted particles.}

\item{converged_fn}{a function that takes a \code{summary} and \code{per_param} input
and generates a logical indicator that the function has converged}

\item{parallel}{parallelise the simulation? If this is set to true then the
simulation step will be parallelised using \code{furrr}. For this to make any
difference it must have been set up with the following:
\code{future::plan(future::multisession, workers = parallel::detectCores()-2)}}

\item{fit}{A S3 \code{abc_fit} object as output by the \code{abc_XXX} functions}

\item{p}{a \code{progressr} progress bar}

\item{seed}{an optional random seed}

\item{knots}{the number of knots to model the CDF with. Optional, and will be
typically inferred from the data size. Small numbers tend to work better if
we expect the distribution to be unimodal.}

\item{allow_continue}{if SMC or adaptive algorithms have not converged after
\code{max_waves} allow the algorithm to interactively prompt the user to continue.}
}
\description{
Common workflow options
}
\concept{workflow}
\keyword{internal}
