% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-transform.R
\name{transform}
\alias{transform}
\title{Generate a distribution from a truncation of another}
\usage{
transform(link, dist, ..., name = NULL)
}
\arguments{
\item{link}{a link function (or name of a link function)}

\item{dist}{distribution(s) as a name, function or a \code{dist_fn} S3 object}

\item{...}{parameters for the underlying distribution if \code{dist} is a name or
function.}

\item{name}{a name for the link function}
}
\value{
a \code{dist_fn} or \code{dist_fn_list} holding the transformed distribution(s)
}
\description{
Generate a distribution from a truncation of another
}
\concept{empirical}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{



t = transform("log","norm")
ps = seq(0,1,0.1)
qs = 0:6
testthat::expect_equal(t$q(ps),qlnorm(ps))
testthat::expect_equal(t$p(qs),plnorm(qs))
testthat::expect_equal(t$d(qs),dlnorm(qs))

t2 = transform("log","norm", 0.4, 0.1)
testthat::expect_equal(t2$q(ps),qlnorm(ps,0.4,0.1))
}\if{html}{\out{</div>}}
}

