% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-truncated.R
\name{truncate}
\alias{truncate}
\title{Generate a distribution from a truncation of another}
\usage{
truncate(dist, x_left, x_right, ...)
}
\arguments{
\item{dist}{a distribution as a name, function or a \code{dist_fn} S3 object}

\item{x_left}{The lower end of the interval or NA for open}

\item{x_right}{The upper end of the interval or NA for open}

\item{...}{parameters for the underlying distribution if \code{dist} is a name or
function.}
}
\value{
a \code{dist_fn} or \code{dist_fn_list} holding the truncated distribution(s)
}
\description{
Generate a distribution from a truncation of another
}
\concept{empirical}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{


shape2_gamma = as.dist_fns(pgamma, shape=2)
g2 = truncate(shape2_gamma, 0.5, 4)

testthat::expect_equal(
  format(g2),
  "trunc(gamma(shape = 2, rate = 1), 0.50 - 4.00); Median (IQR) 1.68 [1.08 — 2.46]"
)

testthat::expect_equal(
  g2$p(0:5),
  c(0, 0.212702667019627, 0.615716430100344, 0.868531239886668, 1, 1)
)

testthat::expect_equal(g2$q(seq(0, 1, 0.2)), c(
  0.5,
  0.971743593113941,
  1.42711359317907,
  1.95304152540128,
  2.6642447272259,
  4
))


g3 = truncate(shape2_gamma, NA, 4)
testthat::expect_equal(
  format(g3),
  "trunc(gamma(shape = 2, rate = 1), 0.00 - 4.00); Median (IQR) 1.54 [0.899 — 2.35]"
)

g4 = truncate(shape2_gamma, 2, NA)
testthat::expect_equal(
  format(g4),
  "trunc(gamma(shape = 2, rate = 1), 2.00 - Inf); Median (IQR) 2.97 [2.42 — 3.87]"
)

}\if{html}{\out{</div>}}
}

