% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-empirical.R
\name{wbw.nrd}
\alias{wbw.nrd}
\title{Weighted bandwidth selector}
\usage{
wbw.nrd(x, w = NULL)
}
\arguments{
\item{x}{data}

\item{w}{weights}
}
\value{
a bandwidth based on weighted data
}
\description{
Weighted bandwidth selector
}
\examples{

x = runif(1000)
w = rgamma(1000,2)

wbw.nrd(x)
wbw.nrd(x,w)

}
\concept{empirical}
